==> Code/20b_EORA_data_inspection.Rmd <==
---
title: "Untitled"
output: pdf_document
editor_options: 
  chunk_output_type: inline
---


# Set up

```{r Paths and packages}
# To clear your environment 
remove(list = ls())

# set work dir
path <- rstudioapi::getSourceEditorContext()$path
dir  <- dirname(rstudioapi::getSourceEditorContext()$path); dir
setwd(dirname(dir))
# setwd('../Data/')
getwd()

source('./Code/_path of data.R')
### packages
source('./Code/_package list.R')

### packages
library(readxl)
library(tidyverse)
library(reshape2) # expand long table to wide one

### plot #####
library(cowplot)
library(ggplot2)
library(ggpubr)
library(RColorBrewer)
## Color Brewer palette
library(viridis)
library(hrbrthemes)
# library(devtools)
library(circlize)  ## for circlize plot
library(Cairo)

today <- format(Sys.time(), "%Y%m%d") ## %H%M%S
```



```{r Theme and font}
### plot settings
unit_ns    <- 'cm'
width_1col <- 8.8   ## 1-column
width_2col <- 18    ## 2-column
font       <- 'sans'     ## "TT Arial"
font_size  <- 7        ##  Nature Sustainability: max = 7; min = 5
theme_ns <- 
  theme_bw()+
  theme(
    # axis.title =element_blank(),
    # axis.text  =element_blank(),
    # axis.ticks =element_blank(),
    # panel.background = element_rect(fill = NA),
    panel.grid.major.x = element_blank(),
    # panel.grid.major = element_blank(),
    # panel.grid.minor = element_line(colour = "red", size = 1),
    legend.background = element_rect(fill="white"),
    legend.key.size = unit(0.15,"cm"),
    text = element_text(size=font_size)
        )
```


#  Data

```{r Ancillary data}
load('./Data/Ancillary_Data_ISO3code_shp.RData') ## iso_eora, pop, gdp, shp, grp

### extended country-iso3 pair list for matching
load('./Data/Ancillary_Data_iso_eora_ex.RData')  ## ## iso_eora_ex

### to match country full_name and abbr 
xls_iso_eora <- "./Data/_eora_190country_iso3_match_update.xlsx"
iso_eora <- readxl::read_excel(path = xls_iso_eora)
```


## Data source

```{r - Dirs}
# dir1 <- './Data/data_02_intermediate/dt02_flows/eora/2022-08-05con/'
# dir2 <- './Data/data_02_intermediate/dt02_flows/eora/2022-08-05pro/'

dir1 <- './Data/data_02_intermediate/dt02_flows/eora/2023-08-16pro/' ## updated on 8/16/2023; only three energy-related files
dir2 <- './Data/data_02_intermediate/dt02_flows/eora/2023-08-17con/' ## updated on 8/17/2023; all files + 1 materiel + 135 shdb
dir2 <- './Data/data_02_intermediate/dt02_flows/eora/2023-08-20con/' ## updated on 8/20/2023; all files
dir2 <- './Data/data_02_intermediate/dt02_flows/eora/2023-08-21con/' ## updated on 8/21/2023; all files


dir_list <- basename(c(dir1, dir2))

country_selected <- c('USA', 'CHN', 'HKG', 'SRB', 'IND', 'IDN', 'JPN', 'DEU', 'FRA', 'RUS', 'CAN', 'BRA', 'PAK', 'VNM') %>% sort()
```


```{r - Indicators}

### indicator file names ----------------------------
# xls.name.eora <- 'YJ_CO2_2015.xlsx'; unit_name <- 'tonne' ## tonne
xls.name.eora <- 'YJ_Water_2015.xlsx'; unit_name <- 'm3'  ## m3
xls.name.eora <- 'YJ_Energytotal_2015.xlsx'; unit_name <- ''         ## PJ = 1000 TJ
# xls.name.eora <- 'YJ_Energy_renewable_2015.xlsx'; unit_name <- ''
# xls.name.eora <- 'YJ_Landuse_forest_2015.xlsx'; unit_name <- 'ha'
# xls.name.eora <- 'YJ_PM25_2015.xlsx'; unit_name <- ''



indicator.name<- gsub('YJ_|\\.xlsx', '', xls.name.eora)

### create file names with path ---------------------
f1 <- paste0(dir1, xls.name.eora);
f2 <- paste0(dir2, xls.name.eora); 
```


```{r - Load data}
### the data
df1 <- readxl::read_excel(path = f1) 

ctrs <- data.frame(
  col = names(df1)[-1], 
  row = df1$Row
)

df1_ <- df1 %>%
  left_join(., iso_eora %>% dplyr::select(Row, iso3_eora), 
            by = c('Row' = 'Row')) %>% ## add ctr code
  dplyr::select(-Row) %>%
  dplyr::rename(exporter_iso3 = iso3_eora) %>%
  dplyr::select(exporter_iso3, everything()) #%>%


## change the full names to iso3 names
names(df1_) <- c("exporter_iso3", df1_$exporter_iso3)
df1.1 <- df1_ %>%
  arrange(!is.na(exporter_iso3)) %>%
  gather(key = 'importer_iso3', value = 'value', 2:ncol(.)) %>%
  arrange(!is.na(importer_iso3)) %>%
  as.data.frame()




df2 <- readxl::read_excel(path = f2) 

ctrs <- data.frame(
  col = names(df2)[-1], 
  row = df2$Row
)

df2_ <- df2 %>%
  left_join(., iso_eora %>% dplyr::select(Row, iso3_eora), 
            by = c('Row' = 'Row')) %>% ## add ctr code
  dplyr::select(-Row) %>%
  dplyr::rename(exporter_iso3 = iso3_eora) %>%
  dplyr::select(exporter_iso3, everything()) #%>%


## change the full names to iso3 names
names(df2_) <- c("exporter_iso3", df2_$exporter_iso3)
df2.1 <- df2_ %>%
  arrange(!is.na(exporter_iso3)) %>%
  gather(key = 'importer_iso3', value = 'value', 2:ncol(.)) %>%
  arrange(!is.na(importer_iso3)) %>%
  dplyr::mutate(value = value) %>%
  as.data.frame()
```






## Compare

```{r - OD}

selected_iso <- c('USA', 'CHN', 'PAK', 'DEU')

d1 <- df1.1 %>%
  dplyr::filter(exporter_iso3 %in% selected_iso) %>%
  dplyr::mutate(value = log10(value)) %>%
  dplyr::rename(value1 = value) %>%
  
  left_join(
    x = .,
    y = df2.1 %>% 
      dplyr::filter(exporter_iso3 %in% selected_iso) %>%
      dplyr::mutate(value = log10(value)) %>%
      dplyr::rename(value2 = value),
    by = c('exporter_iso3', 'importer_iso3')
  ) %>%
  ### remove the self-to-self trade (too big values)
  # dplyr::filter(exporter_iso3 != importer_iso3) %>%
  dplyr::mutate(dif_abs = abs(value2 - value1))


# Filter the data to select only the points to annotate
selected_data <- rbind(
  d1 %>% group_by(exporter_iso3) %>% slice_max(prop = 0.02, order_by = value1), 
  d1 %>% group_by(exporter_iso3) %>% slice_min(prop = 0.02, order_by = value1),
  d1 %>% group_by(exporter_iso3) %>% slice_max(prop = 0.03, order_by = dif_abs),
  )

library("ggrepel") 
d1 %>%
  ggplot(data = ., aes(x = value1, y = value2, color = exporter_iso3)) +
  geom_point(alpha = 0.5, show.legend = F) +
  facet_wrap(~exporter_iso3, scales = 'free') +
  
  # geom_text(data = selected_data, aes(label = importer_iso3), nudge_x = 0.1, nudge_y = 0.1) + 
  # geom_label(data = selected_data, aes(label = importer_iso3), vjust = 1, nudge_y = 2) +
  geom_text_repel(
    data = selected_data, aes(label = importer_iso3), 
    color = 'blue', alpha = .8,
    min.segment.length = 0, # draw all line segments
    point.padding = 0, # additional padding around each point
    box.padding = 0.5, # additional padding around each text label
    max.overlaps = Inf) + 

  geom_abline(slope = 1, color = 'red', alpha = .8) +
  xlab(dir_list[1]) +
  ylab(dir_list[2]) +
  ggtitle(indicator.name) +
  theme_bw()

### save plot
fname <-paste0(dir.fig, 'eora_figures/', 'data_inspection_Trade_', indicator.name, '_', 
               paste(dir_list, collapse = '_'), '_',today, '.png'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = 9, height = 9, units = 'in', dpi = 200)
```


### heatmap
```{r}
### plot to see
myPalette <- colorRampPalette((brewer.pal(9, "Blues")));# myPalette; rev
n1 <- 9
colors <- myPalette(n1)[0:(n1)]; colors
### generate a vector of n contiguous colors using the functions 
### rainbow(n), heat.colors(n1), terrain.colors(n), topo.colors(n), and cm.colors(n2).
n2 <- 100
colors <- topo.colors(n2) %>% rev()
colors <- heat.colors(n2) %>% rev()
colors <- terrain.colors(n2) %>% rev() 

space_size = 0.3

ht1 <- df1.1 %>%
  dplyr::filter(exporter_iso3 %in% country_selected,
                importer_iso3 %in% country_selected,) %>%
    ### transform the data to be comparable
    dplyr::mutate(value = log10(value)) %>%
    ### remove the self-to-self trade (too big values)
    dplyr::filter(exporter_iso3 != importer_iso3) %>%
    ggplot(data = ., aes(x = importer_iso3, y = exporter_iso3)) + 
    geom_tile(aes(fill = value), colour = "white", 
              size = space_size, ## the space between boxes
              show.legend = T) + 
    scale_fill_gradientn(colours = colors) +
    # theme_bw() +
    theme_ns +
    xlab('destination nation') + ylab('origin nation') +
    labs(fill = paste0(indicator.name, ': log10(', unit_name, ')')) +
    theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.3))


ht2 <- df2.1 %>%
  dplyr::filter(exporter_iso3 %in% country_selected,
                importer_iso3 %in% country_selected,) %>%
  ### transform the data to be comparable
  dplyr::mutate(value = log10(value)) %>%
  ### remove the self-to-self trade (too big values)
  dplyr::filter(exporter_iso3 != importer_iso3) %>%
  ggplot(data = ., aes(x = importer_iso3, y = exporter_iso3)) + 
  geom_tile(aes(fill = value), colour = "white", 
            size = space_size, ## the space between boxes
            show.legend = T) + 
  scale_fill_gradientn(colours = colors) +
  # theme_bw() +
  theme_ns +
  xlab('destination nation') + ylab('origin nation') +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.3))

ggarrange(ht1, ht2, ncol = 2, labels = dir_list, common.legend = T, vjust = 0, legend = 'top')

### save plot
fname <-paste0(dir.fig, 'eora_figures/', 'data_inspection_heatmap_', indicator.name, 
               '_', paste(dir_list, collapse = '_'),  '_', today, '.png'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = width_2col, height = width_2col/2, units = unit_ns, dpi = 300)
```
