==> Code/21_MRIO_data_clean.Rmd <==
---
title: "MRIO data cleaning"
author: "Yingjie Li"
date: "2021-03-01"
output: pdf_document
editor_options: 
  chunk_output_type: inline
---


  This script aims to 
  
  1. Combine the 4-year files for each footprint indicator in to one single dataframe. 
  
  2. Format the dataframe (e.g., rename, match country ISO3 code) and save it to `eora_cleaned` folder.
  
  3. Use heatmap to quick check the data quanlity. 
  
# Set up
```{r Paths and packages}
knitr::opts_chunk$set(echo = TRUE)

### To clear your environment 
remove(list = ls())

### set work dir
path <- rstudioapi::getSourceEditorContext()$path
dir  <- dirname(rstudioapi::getSourceEditorContext()$path); dir
setwd(dir)
setwd('..') # set directory by one folder up
getwd()

### data path 
source('./Code/_path of data.R')


### packages
source('./Code/_package list.R')
```



```{r Theme and font}
theme_set(theme_bw())


### plot settings
unit_ns    <- 'cm'
width_1col <- 8.8   ## 1-column
width_2col <- 18    ## 2-column
font       <- 'sans'     ## "TT Arial"
font_size  <- 8          ##  Nature Sustainability: max = 7; min = 5
theme_ns <- 
  theme_bw()+
  theme(
    # axis.title =element_blank(),
    # axis.text  =element_blank(),
    # axis.ticks =element_blank(),
    # panel.background = element_rect(fill = NA),
    panel.grid.major.x = element_blank(),
    # panel.grid.major = element_blank(),
    # panel.grid.minor = element_line(colour = "red", size = 1),
    legend.background = element_rect(fill="transparent"),
    legend.key.size = unit(0.15,"cm"),
    text = element_text(size=font_size)
        )
```


```{r Ancillary data}
load('./Data/Ancillary_Data_ISO3code_shp.RData') ## iso_eora, pop, gdp, shp, grp

### extended country-iso3 pair list for matching
load('./Data/Ancillary_Data_iso_eora_ex.RData')  ## ## iso_eora_ex


### use the Eora data as the template 
f <- paste0('./Data/data_02_intermediate/dt02_flows/', 'template_eora.RData')
load(f) # `temp_eora`


### to match country full_name and abbr 
xls_iso_eora <- "./Data/_eora_190country_iso3_match_update.xlsx"
iso_eora <- readxl::read_excel(path = xls_iso_eora)
```




#  Data

## Eora data Batch 

  Batch read in and format as trade matrix among countries, for 2000, 2005, 2010, and 2015.
  
  Combine the 4-year data for each ft indicator into one single file, and save to `eora_cleaned` folder. 
  

```{r - data version ❗❗}
### to choose data version by modifying `dir.eora` in `_path of data.R`


### to list all the input data files
library(stringr)
pattern <- '^YJ_'
xls.eora.ls <- list.files(path = dir.eora, pattern = pattern, full.names = T); 
xls.ls <- basename(xls.eora.ls); 


### to obtain the unique list of all the footprint indicator names
### - old list
xls.eora.ls.old <- list.files(path = paste0(dirname(dir.eora), '/2021-08-05'), pattern = pattern, full.names = T);
xls.ls.old <- basename(xls.eora.ls.old);
inds <- substr(x = xls.ls.old, start = 4, stop = max(nchar(xls.ls.old))); 
inds <- str_sub(string = inds, end = -10);
inds <- gsub(pattern = '_$', '', x = inds);
inds <- unique(inds); 
inds.old <- inds
cat(paste0('\n The total number of Eora indicators: ', length(inds.old))) ## 54


### - new list (update on 2022-08-01)
inds <- substr(x = xls.ls, start = 4, stop = max(nchar(xls.ls))); 
inds <- str_sub(string = inds, end = -10);           ## remove "_YYYY.xlsx"
inds <- gsub(pattern = '_$|20220801', '', x = inds); ## fix format issue in SHDB' names
inds <- gsub(pattern = 'SHDB_', 'SHDB', x = inds);   ## fix format issue in SHDB' names
inds <- unique(inds); 
cat(paste0('\n The total number of Eora indicators: ', length(inds)), '\n\n')

### - difference
setdiff(inds.old, inds)
setdiff(inds, inds.old)


### save this indicator list as a file
inds_df <- data.frame(indicator = inds)
writexl::write_xlsx(x = inds_df, path = './Data/_eora_indicators_io_list.xlsx')
```
  


```{r - loop read and combine xlsx ❗, include=FALSE}
# pattern <- '^YJ_CO2'
n_year_check <- 4 ## if 4 years data, to make sure no more files will be matched. 
n_year_check <- 1
f <- paste0(dir.flowScenario, 'para_n_year_check.RData'); f
save(n_year_check, file = f)

##' remove SHDB files from the list
inds_nonSHDB <- inds[grep("SHDB", inds, invert = T, ignore.case = TRUE)]

for (i in inds_nonSHDB) {
  # print(i)
  iname = i
  pattern <- paste0('^YJ_', i, '_20'); #print(pattern)
  xls.eora.ls <- list.files(path = dir.eora, pattern = pattern, full.names = T); xls.eora.ls
  
  ###' in case the pattern would match too many files, 
  ###' we use this `if else` function to pop out a warning message if anything goes wrong. 
  if(length(xls.eora.ls) > n_year_check | length(xls.eora.ls) < (n_year_check-1) ) {
    # print('Match error!') 
    stop('Match error!')
  } else {
    
    ### a null data frame
    df.eora.x.all <- data.frame()
    for (xls in xls.eora.ls) {
      # print(xls)
  
      ### Extracting the last n characters from a string in R, and get the year number
      yr <- str_sub(xls, -9, -6)
      df.eora  <- readxl::read_excel(path = xls)
      df.eora1 <- df.eora %>%
        left_join(., 
                  iso_eora %>% dplyr::select(Row, iso3_eora), 
                  by = c('Row' = 'Row')) %>% ## add ctr code
        arrange(!is.na(iso3_eora)) %>%
        as.data.frame()
  
      ### rename the cols
      ctr_ls_eora <- unique(df.eora1$iso3_eora); ctr_ls_eora
      names(df.eora1) <- c('ctr', ctr_ls_eora, 'iso3')
  
      ### reorder the col names
      col_name_ctr <- names(df.eora1)[-c(1, ncol(df.eora1))];  ## remove "ctr" in the 1st col and "iso3" in the last
      # print(col_name_ctr)
      print(paste0('Number of countries: ', length(col_name_ctr)))
  
      df.eora.yr <- df.eora1 %>%
        dplyr::mutate(year = yr) %>%
        dplyr::select(year, ctr, iso3, all_of(col_name_ctr))
  
      ### append all 4 years data in one table
      df.eora.x.all <- rbind(df.eora.x.all, df.eora.yr);
  
    }
  
    print(paste0('Total number of rows for the combine data set: ', nrow(df.eora.x.all)))
  
    ### save as xlsx
    fname <- paste0(dir.eora_cleaned, iname, '.xlsx'); cat('\n', fname, '\n\n')
    writexl::write_xlsx(x = df.eora.x.all, path = fname)
  }
  
}

```




```{r - loop detailed SHDB}
dir.shdb.detail <- gsub('135', '135_renamed', dir.shdb)
pattern <- '^YJ_SHDB'
inds_ls <- list.files(path = dir.shdb.detail, pattern = pattern, full.names = T);
  
for (xls in inds_ls) {
  
  ### Extracting the last n characters from a string in R, and get the year number
  yr <- str_sub(xls, -9, -6)
  df.eora  <- readxl::read_excel(path = xls)
  df.eora1 <- df.eora %>%
    left_join(., 
              iso_eora %>% dplyr::select(Row, iso3_eora), 
              by = c('Row' = 'Row')) %>% ## add ctr code
    arrange(!is.na(iso3_eora)) %>%
    as.data.frame()

  ### rename the cols
  ctr_ls_eora <- unique(df.eora1$iso3_eora); ctr_ls_eora
  names(df.eora1)
  names(df.eora1) <- c('ctr', ctr_ls_eora, 'iso3')

  ### reorder the col names
  col_name_ctr <- names(df.eora1)[-c(1, ncol(df.eora1))];  ## remove "ctr" in the 1st col and "iso3" in the last
  # print(col_name_ctr)
  print(paste0('Number of countries: ', length(col_name_ctr)))

  df.eora.yr <- df.eora1 %>%
    dplyr::mutate(year = yr) %>%
    dplyr::select(year, ctr, iso3, all_of(col_name_ctr))

  ### append all 4 years data in one table
  # df.eora.x.all <- rbind(df.eora.x.all, df.eora.yr);
  df.eora.x.all <- df.eora.yr

  print(paste0('Total number of rows for the combine data set: ', nrow(df.eora.x.all)))

  ### save as xlsx
  iname <- basename(xls) %>% gsub('YJ_', '', x=.) %>% gsub(paste0('_',yr), '', x=.)
  fname <- paste0(dir.eora_cleaned, iname); print(fname); cat('\n\n')
  writexl::write_xlsx(x = df.eora.x.all, path = fname)
}

```



  SHDB data has only 3 years (i.e., 2014, 2015, and 2019). To make all the data files have the same number of rows for easy data processing, we need to list `NA` data for 2000. 
  
```{r - format SHDB data}

xls  <- paste0(dir.eora_cleaned, "SHDB1.xlsx"); xls
temp <- readxl::read_excel(path = xls) %>%
    dplyr::filter(year == '2019') 


### if use 4 years of data, we need to add year 2000 with NA values
if (n_year_check == 4) {
  
  ### 1. to create NA data for 2000 
  add_na_2000 <- temp %>%
    dplyr::mutate_if(is.numeric, funs(ifelse(!is.na(.), NA, .))) %>%
    dplyr::mutate(year = '2000')
  # str(add_na_2000)
  
  
  ### 2. format all the SHDB data by adding SHDB in 2000 with NA 
  xls.ls <- list.files(path = dir.eora_cleaned, pattern = 'xlsx$', full.names = T); #xls.ls
  xls.ls.SHDB <- str_subset(xls.ls, pattern="SHDB"); xls.ls.SHDB
  
  for (xls in xls.ls.SHDB) {
    print(xls)
    df  <- readxl::read_excel(path = xls)
    yrs <- unique(df$year); yrs
    
    ## if it does not have data for 2000, then add NA data; if includes, then skip and do nothing
    if(!is.element('2000', yrs)) {           ## !'2000' %in% yrs
      print('to add NA values to the year of 2000') 
      df_new <- rbind(add_na_2000, df)
      writexl::write_xlsx(x = df_new, path = xls)
    } else{
      print('skip')
    }
  }


  ### if we only use 1 year data, then skip this formatting step
} else {
  temp <- readxl::read_excel(path = xls) 
  cat('Only', length(unique(temp$year)), 'year data is included here:\t', unique(temp$year), '\n')
}

```





## Other data sources

###  - arms trade
```{r, eval=FALSE, include=FALSE}
xls <- list.files(path = dir.arm, pattern = '_TIV_arms_trade_', full.names = T); xls
df <- readxl::read_excel(path = xls) #%>%
  ### format as the same as EORA data
  # dplyr::mutate(ctr = NA) %>%
  # dplyr::select(year, ctr, iso3, everything())
## get the indicator name
ind <- 'arms_trade'

### Format the data in the same way as Eora cleaned ------------------------------------------------
df_formatted <- df %>%
  dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) %>%
  gather(key = 'to', value = 'value', 3:ncol(.)) %>%
  merge(x = temp_eora, y = ., by.x = c('year', 'iso3', 'to'), by.y = c('year', 'iso3', 'to'), all.x = T)

### To generate the same format as Eora in Matrix
df_formatted <- df_formatted %>%
  ungroup() %>%
  arrange(year, iso3, to) %>%
  dplyr::select(year, ctr, everything()) %>%
  spread(key = to, value = value) %>%
  arrange(year, iso3)


### save to the `eora_cleaned` folder --------------------------------------------------------------
df_formatted %>%
  group_by(year) %>% tally()
fname <- paste0(dir.eora_cleaned, ind, '.xlsx'); fname
writexl::write_xlsx(x = df_formatted, path = fname)



### Calculate net_imports --------------------------------------------------------------------------
# net_imports <- data.frame()
# net_in <- func_net_import(df = df)
# net_imports <- rbind(net_imports, cbind(ind = ind, net_in)) %>%
#   arrange(iso3, year)
# 
# 
# ### check the data
# sort(unique(net_imports$year))
# unique(net_imports$ind)
# length(unique(net_imports$ind))
# 
# 
# net_imports3 <- net_imports %>%
#   dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) 
```




###  - human_trafficking
```{r, eval=FALSE, include=FALSE}

xls <- list.files(path = dir.traffick, pattern = '_human_trafficking', full.names = T); xls
df <- readxl::read_excel(path = xls) %>%
  ### format as the same as EORA data
  gather(key = 'to', value = 'value', 3:ncol(.)) %>%
  pivot_wider(names_from = year, values_from = value) %>%
  dplyr::mutate( `2000` = `2002`) %>%
  # dplyr::mutate(ctr = NA) %>%
  dplyr::select(iso3, everything()) %>%
  gather(key = 'year', value = 'value', 3:ncol(.)) %>%
  pivot_wider(names_from = to, values_from = value) %>%
  # arrange(iso3, year) %>%
  # dplyr::select(year, ctr, iso3, everything())
  as.data.frame()

## get the indicator name
ind <- 'human_trafficking'


### save to the `eora_cleaned` folder
df_4yr <- df %>%
  ungroup() %>% 
  arrange(year, iso3) %>%
  group_by(year, iso3) %>% ## ??? too few data, should use data in nearby years to fill the gap
  dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) 
df_4yr %>%
  group_by(year) %>% tally() ## ??? need to add all countries to the table, with NA



### Format the data in the same way as Eora cleaned ------------------------------------------------
df_formatted <- df %>%
  dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) %>%
  gather(key = 'to', value = 'value', 3:ncol(.)) %>%
  merge(x = temp_eora, y = ., by.x = c('year', 'iso3', 'to'), by.y = c('year', 'iso3', 'to'), all.x = T)

### To generate the same format as Eora in Matrix
df_formatted <- df_formatted %>%
  ungroup() %>%
  arrange(year, iso3, to) %>%
  dplyr::select(year, ctr, everything()) %>%
  spread(key = to, value = value) %>%
  arrange(year, iso3)


### save to the `eora_cleaned` folder --------------------------------------------------------------
df_formatted %>%
  group_by(year) %>% tally()
fname <- paste0(dir.eora_cleaned, ind, '.xlsx'); fname
writexl::write_xlsx(x = df_formatted, path = fname)



### Calculate net_imports --------------------------------------------------------------------------
# net_imports <- data.frame()
# net_in <- func_net_import(df = df)
# net_imports <- rbind(net_imports, cbind(ind = ind, net_in)) %>%
#   arrange(iso3, year)
# 
# ### check the data
# sort(unique(net_imports$year))
# unique(net_imports$ind)
# length(unique(net_imports$ind))
# 
# net_imports4 <- net_imports %>%
#   dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) 
```


### - Material footprint

  See data process in `10_country_profileData_prepare.Rmd`
```{r}

```


### - peacekeeping?
```{r }
### ".\Data\data_01_raw\Arm trade_Peacekeeping\UN_peacekeeping.Rmd"


# xls <- list.files(path = dir.traffick, pattern = '_human_trafficking', full.names = T); xls
# df <- readxl::read_excel(path = xls) %>%
#   ### format as the same as EORA data
#   gather(key = iso3.y, value = value, 3:ncol(.)) %>%
#   pivot_wider(names_from = year, values_from = value) %>%
#   dplyr::mutate(ctr = NA, `2000` = `2002`) %>%
#   dplyr::select(ctr, iso3, everything()) %>%
#   gather(key = year, value = value, 4:ncol(.)) %>%
#   pivot_wider(names_from = iso3.y, values_from = value) %>%
#   arrange(iso3, year) %>%
#   dplyr::select(year, ctr, iso3, everything())
# ## get the indicator name
# ind <- 'human_trafficking'

```


###  - student_flow
```{r, eval=FALSE, include=FALSE}
### get the indicator name
ind <- 'student_flow'

# dir.student_flow <- './Data/data_01_raw/International Student flows/' ## have pre-setted
dir.student_flow

xls <- list.files(path = dir.student_flow, pattern = 'output_student_flows_ALL_wide_', full.names = T); xls
df <- readxl::read_excel(path = xls) %>%
  ### format as the same as EORA data
  # dplyr::mutate(ctr = NA) %>%
  dplyr::rename(iso3 = send_iso) %>%
  dplyr::select(year, iso3, everything())



### check the number of countries in the data 
df_4yr <- df %>%
  ungroup() %>% 
  arrange(year, iso3) %>%
  group_by(year, iso3) %>% ## ??? if too few data, should use data in nearby years to fill the gap
  dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) 
df_4yr %>%
  group_by(year) %>% tally() ## ??? need to only include the 189 countries

### Format the data in the same way as Eora cleaned ------------------------------------------------
df_formatted <- df %>%
  dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) %>%
  gather(key = 'to', value = 'value', 3:ncol(.)) %>%
  merge(x = temp_eora, y = ., by.x = c('year', 'iso3', 'to'), by.y = c('year', 'iso3', 'to'), all.x = T)

### To generate the same format as Eora in Matrix
df_formatted <- df_formatted %>%
  ungroup() %>%
  arrange(year, iso3, to) %>%
  dplyr::select(year, ctr, everything()) %>%
  spread(key = to, value = value) %>%
  arrange(year, iso3)


### save to the `eora_cleaned` folder --------------------------------------------------------------
df_formatted %>%
  group_by(year) %>% tally()
fname <- paste0(dir.eora_cleaned, ind, '.xlsx'); fname
writexl::write_xlsx(x = df_formatted, path = fname)


### Calculate net_imports --------------------------------------------------------------------------
# net_imports <- data.frame()
# net_in <- func_net_import(df = df)
# net_imports <- rbind(net_imports, cbind(ind = ind, net_in)) %>%
#   arrange(iso3, year)
# 
# ### check the data
# sort(unique(net_imports$year))
# unique(net_imports$ind)
# length(unique(net_imports$ind))
# 
# net_imports5 <- net_imports %>%
#   dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) 
```



###  - TradeMatrix_marine_capture
  Refer to the report "OECD FishPEM Model Capture Fisheries Trade Data" by Dr. Jessica Gephart. 
```{r, eval=FALSE, include=FALSE}
### get the indicator name
ind <- 'TradeMatrix_marine_capture'

# dir.fish <- './Data/data_01_raw/FAOSTAT/Jessica Gephart/' ## pre-set 
dir.fish

# f <- 'marine_capture_trade_20211214.csv'        ## 1st version 2021-12-14
f <- 'marine_capture_trade_5yrave_20220107.csv'   ## 2nd version 2022-01-07
file <- paste0(dir.fish, f)


df0 <- readr::read_csv(file = file, show_col_types = FALSE) %>%
  # dplyr::distinct(exporter_iso3c, importer_iso3c, .keep_all = T) %>%
  dplyr::select(exporter_iso3c, importer_iso3c, `q.live`) %>%
  as.data.frame()


df1 <- df0 %>%
  dplyr::mutate(year = 2015) %>%
  spread(key = importer_iso3c, value = `q.live`)

df1 %>%
  group_by(year) %>% tally() ## 220 countries are in the raw data, but we need to only include the 189 countries
```



```{r - update data}
f <- 'artis_marine_capture_20230915.csv'   ## 3rd version 2023-09-15
file <- paste0(dir.fish, f)

df02 <- readr::read_csv(file = file, show_col_types = FALSE) %>%
  # dplyr::distinct(exporter_iso3c, importer_iso3c, .keep_all = T) %>%
  as.data.frame()

## compare with last version 
df0.compare <- merge(df0, df02, by=c('exporter_iso3c', 'importer_iso3c'), all = T)
df0.compare %>%
  ggplot(aes(x = q.live, y = live_weight_t)) +
  geom_point(alpha = 0.3) +
  geom_abline(slope = 1, color = 'red') +
  scale_x_log10(breaks = trans_breaks("log10", function(x) 10^x),
                  labels = trans_format("log10", math_format(10^.x))) +
  scale_y_log10(breaks = trans_breaks("log10", function(x) 10^x),
                labels = trans_format("log10", math_format(10^.x))) +
  xlab('Old data - 20220107') + ylab('New data - 20230915') +
  theme_bw()

df2 <- df02 %>%
  dplyr::mutate(year = 2015) %>%
  dplyr::select(year, exporter_iso3c, importer_iso3c, `live_weight_t`) %>%
  spread(key = importer_iso3c, value = `live_weight_t`)

df2 %>% group_by(year) %>% tally()
```


```{r, eval=FALSE, include=FALSE}

### choose data version to use for further analysis
df <- df2


### Format the data in the same way as Eora cleaned
df_formatted <- df %>%
  dplyr::rename(iso3 = exporter_iso3c) %>%
  dplyr::filter(year %in% c(2000, 2005, 2010, 2015)) %>%
  gather(key = 'to', value = 'value', 3:ncol(.)) %>%
  merge(x = temp_eora, y = ., by.x = c('year', 'iso3', 'to'), by.y = c('year', 'iso3', 'to'), all.x = T)

### To generate the same format as Eora in Matrix
df_formatted <- df_formatted %>%
  ungroup() %>%
  arrange(year, iso3, to) %>%
  dplyr::select(year, ctr, everything()) %>%
  spread(key = 'to', value = 'value') %>%
  arrange(year, iso3)


### save to the `eora_cleaned` folder
df_formatted %>%
  group_by(year) %>% tally()
fname <- paste0(dir.eora_cleaned, ind, '.xlsx'); fname
writexl::write_xlsx(x = df_formatted, path = fname)

```



###  - TradeMatrix_Cereals
```{r}

# See "Cereal trade - maxtrix" section in `10_country_profileData_prepare.Rmd`

```




### - TradeMatrix_Pesticides_Hzd
```{r }

# see `11_Comtrade_BACI_Pesticides.Rmd`

```



### - NCD death embodied in trade matrix
```{r}

# See `NCDs_matrix.R`

# source('./Data/data_01_raw/Health/NCDs_MinGonChung/NCDs_matrix.R')

# ind <- "TradeMatrix_NCD_death_number"
```




### - FDI - IMF

  See `FDI_data_prep.Rmd` in "./Data/data_01_raw/FDI/"
  
```{r}
### get the indicator name
ind <- 'FDI_matrix'
yrs <- c(2000, 2005, 2010, 2015)
dir.i <- './Data/data_01_raw/FDI/IMF/'

fname <- paste0(dir.eora_cleaned, ind, '.xlsx'); fname
df_formatted <- readxl::read_excel(fname)
```


### - ODA OECD

  See `ODA_data_prep.Rmd` in "./Data/data_01_raw/OECD_ODA/"

```{r}
ind <- 'ODA_OOF_total_matrix'
ind <- 'ODA_Infrastructure_matrix'
ind <- 'ODA_Agriculture_matrix'
ind <- 'ODA_Education_matrix'
ind <- 'ODA_Energy_matrix'
ind <- 'ODA_Food_matrix'
ind <- 'ODA_Water_matrix'
ind <- 'ODA_Biodiversity_matrix'
ind <- 'ODA_total_disbursements_matrix'
ind <- 'ODA_Health_matrix'

yrs <- c(2000, 2005, 2010, 2015)
dir.i <- './Data/data_01_raw/OECD_ODA/'

fname <- paste0(dir.eora_cleaned, ind, '.xlsx'); fname
df_formatted <- readxl::read_excel(fname)
```



### - Portfolio Investment - IMF-CPIS

  See `CPIS_data_prep.Rmd` in "./Data/data_01_raw/IMF_CPIS/"
  
```{r}
### get the indicator name
ind <- 'CPIS_matrix'
yrs <- c(2000, 2005, 2010, 2015)
dir.i <- './Data/data_01_raw/IMF_CPIS/'

fname <- paste0(dir.eora_cleaned, ind, '.xlsx'); fname
df_formatted <- readxl::read_excel(fname)
```


### - Bilateral Remittance Matrix - WB

  See `Remittance_data_prep.Rmd` in "./Data/data_01_raw/WB_Remittance/"
  
```{r}
### get the indicator name
ind <- 'Remittance'
yrs <- c(2000, 2005, 2010, 2015)

fname <- paste0(dir.eora_cleaned, ind, '_matrix.xlsx'); fname
df_formatted <- readxl::read_excel(fname)
```



### - Inventor migration

  See `Inventor migration_data_prep.Rmd` in './Data/data_01_raw/Inventor migration files/'
  
```{r}
### get the indicator name
ind <- 'Inventor migration'
yrs <- c(2000, 2005, 2010, 2015)

fname <- paste0(dir.eora_cleaned, ind, '_matrix.xlsx'); fname
df_formatted <- readxl::read_excel(fname)
```


### - High-skilled immigrants

  See `Immigrants_data_prep.Rmd` in './Data/data_01_raw/OECD_Immigrants/'
  
```{r}
### get the indicator name
ind <- 'High-skilled immigrants'
yrs <- c(2000, 2005, 2010, 2015)

fname <- paste0(dir.eora_cleaned, ind, '_matrix.xlsx'); fname
df_formatted <- readxl::read_excel(fname)
```


### - Refugee - UNHCR

  See `Refugee_data_prep.Rmd` in "./Data/data_01_raw/IMF_CPIS/"
  
```{r}
### get the indicator name
ind <- 'Refugee'
yrs <- c(2000, 2005, 2010, 2015)
dir.i <- './Data/data_01_raw/UNHCR_Refugee Data Finder/'

fname <- paste0(dir.eora_cleaned, ind, '_matrix.xlsx'); fname
df_formatted <- readxl::read_excel(fname)
```




## Heatmap - help check data

  A heat map can more vivid to check the data quality. Here we took data in 2015 as an example. 
  
```{r}

xls.ls <- list.files(path = dir.eora_cleaned, pattern = 'xlsx$', full.names = T); xls.ls

yr <- 2015

for (xls in xls.ls) {
  # print(xls)
  df <- readxl::read_excel(path = xls) %>%
    dplyr::filter(year == yr) %>%
    as.data.frame()
  # yrs <- unique(df$year); print(yrs)
  iname <- gsub('.xlsx', '', basename(xls))
  
  
  space_size = 0.3
  
  dfht <- df %>%
    gather(key = 'receive', value = 'value', 4:ncol(.)) %>%
    arrange(iso3, receive) %>%
    ### transform the data to be comparable
    dplyr::mutate(value = log10(value)) %>%
    ### remove the self-to-self trade (too big values)
    # dplyr::filter(iso3 != receive) %>%
    ggplot(data = ., aes(x = receive, y = iso3)) + 
    geom_tile(aes(fill = value), colour = "white", 
              # na.rm = T, 
              size = space_size, ## the space between boxes
              show.legend = T) + 
    scale_fill_gradientn(colours = topo.colors(100), na.value = "gray50") +
    # theme_bw() +
    theme_ns +
    xlab('destination nation') + ylab('origin nation') +
    theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.3))
  # dfht
  
  ### save plot
  fname <-paste0(dir.fig, 'footprint_heatmap/FT_HT_', iname, '_', yr, '_', postfix, '_', today, '.png'); print(fname)
  ggsave(filename = fname, plot = dfht, width = 16, height = 16, units = 'in', dpi = 300)
  
}
```


  Rename files. 
```{r eval=FALSE, include=FALSE}
name_old <- list.files(path = paste0(dir.fig, 'footprint_heatmap/'), pattern="20220812.png", full.names = T)
name_old
name_new <- gsub('20220812.png', '20220812con.png', name_old)
name_new
# file.rename(from = name_old, to = name_new)
```
