==> Code/22_20_TNI_dst_ner.Rmd <==
---
title: "Spillover Analysis"
author: "Yingjie Li"
date: "2021-09-23"
output: pdf_document
editor_options: 
  chunk_output_type: inline
---


  This script aims to 
  
  1. Read the combined EORA indicator data (4 years data in one table), and calculate `net imports` for each indicator, each year, and each country. 
  
  2. Calculate `spillover score` and `spillover index`, according to SDSN's approach. 
  
  
# Set up

```{r}

### To clear your environment 
remove(list = ls())

### packages
source('./Code/_package list.R')
source('./Code/func_net_import_direction.R')

### data path
source('./Code/_path of data.R')

today <- format(Sys.time(), "%Y%m%d"); today
```


#  Data 

## Neighbor matrix

```{r}
### load spatial relation: nearby or distant
ner_dst_matrix <- './Data/_neighbor_matrix/_input_ctr_matrix_neighbor.xlsx'
ner.matrix <- read_excel(path = ner_dst_matrix) %>% dplyr::select(-Country, -Neighbors) %>% column_to_rownames(var = "o")
diag(ner.matrix) <- 1
dst.matrix <- 1 - ner.matrix               
not.matrix <- (ner.matrix + dst.matrix)

## `ner.matrix`: nearby = 1, distant = 0
## `dst.matrix`: nearby = 0, distant = 1


### how many nearby and distant pairs?
ner_stat <- ner.matrix %>%
  rownames_to_column(var = 'o') %>%
  gather(key = 'd', value = 'scenario', 2:ncol(.)) %>%
  group_by(scenario) %>%
  tally() %>%
  dplyr::mutate(total = sum(n, na.rm = T),
                percent = scales::percent(n/total, accuracy = 0.1))
ner_stat
ner_stat$n/sum(ner_stat$n) * 100

### --> 96.9% distant pairs; 3.1% nearby pairs



### how many nearby or distant partners each country has? --------------------------------
ner_stat_each <- ner.matrix %>%
  dplyr::mutate(n_nearby = rowSums(.),
                n_near_pct = n_nearby/189*100) %>% ## % are nearby
  dplyr::select(n_nearby, n_near_pct) %>%
  as.data.frame()

hist(ner_stat_each$n_nearby)
boxplot(ner_stat_each$n_nearby)
summary(ner_stat_each$n_nearby)
  #  Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  # 1.000   4.000   5.000   5.852   7.000  24.000 


### --> for distant pairs
dst_stat_each <- dst.matrix %>%
  dplyr::mutate(n     = rowSums(.),
                n_pct = n/189*100) %>% ## % are nearby
  dplyr::select(n, n_pct) %>%
  as.data.frame()
hist(dst_stat_each$n)
boxplot(dst_stat_each$n)
summary(dst_stat_each$n)
  #  Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  # 165.0   182.0   184.0   183.1   185.0   188.0 
```




## Net imports and TNI Score by senarios

  Only indicators with decided *direction* were included in the calculation, while assign *NA* to those with un-decided direction. 
  - Need to update the `direction` table!
  
```{r }
### get the name list of all the spillovers and manually add `direction` info for further normalizing the values

xls.ls <- list.files(path = dir.eora_cleaned, pattern = 'xlsx$', full.names = T); 

spillover_name_list <- data.frame(ind = gsub('.xlsx', '', basename(xls.ls))) %>%
  # dplyr::mutate(shdb = str_detect(ind, pattern="SHDB")) %>%
  dplyr::mutate(direction = '') %>%  ## to determine if it is positive or negative impact
  as.data.frame()


### loop the normalization based on `direction` ------------------------------------------

f <- paste0('./Data/data_02_intermediate/dt02_flows/', 'direction.RData')
load(f) ## `direction`, `direction_clean`

direction_use <- direction %>%
  dplyr::mutate(direction = as.numeric(direction)) %>%
  dplyr::filter(!is.na(direction))
```




**Warning !!!**

  * In the scenario of `only nearby` interaction, we need to use the 'burden' from `distant` interactions (Be careful here!) 
  * `distant only` scenario = no nearby  trade: to evaluate *nearby*  impact --> use nearby  footprints data
  * `nearby only`  scenario = no distant trade: to evaluate *distant* impact --> use distant footprints data

```{r - distant only scenario --> nearby impact, include=FALSE}

scenario <- 'fromNer'
matrix   <- ner.matrix


### put all spillover indicators in one table 
net_imports <- data.frame()


### this for loop can be used in other scenarios 
for (i in 1:nrow(direction_use)) {
  # print(i)
  ind <- direction_use[i, 'ind']       %>% unlist() %>% paste(sep = '', collapse = ''); 
  drt <- direction_use[i, 'direction'] %>% unlist() %>% paste(sep = '', collapse = '') %>% as.numeric()
  in_out <- direction_use[i, 'Impact_direction'] %>% unlist() %>% paste(sep = '', collapse = '')
  print(ind)
  print(drt)
  
  
  ### if the direction is specified, read data; if not, skip. 
  if (!is.na(drt)) {
    xls <- paste0(dir.eora_cleaned, ind, '.xlsx'); print(basename(xls))
    df  <- readxl::read_excel(path = xls)  %>% 
      gather(key = 'd', value = 'value', 4:ncol(.)) %>%
      as.data.frame()
    
    df_wide <- df %>%
      dplyr::group_by(year) %>%
      arrange(year, iso3, d) %>%
      spread(key = 'd', value = 'value') %>%
      arrange(year, iso3) %>%
      as.data.frame()
      
    ### --> `df` under different scenarios
    mat <- matrix %>%
      rownames_to_column(var = 'o') %>%
      gather(key = 'd', value = 'scenario', 2:ncol(.))
    
    df_scenario <- merge(df, mat, by.x = c('iso3', 'd'), by.y = c('o', 'd'), all.x = T) %>%
      dplyr::mutate(value = value * scenario) %>%
      dplyr::select(-scenario) %>%
      spread(key = 'd', value = 'value') %>%
      arrange(year, iso3) %>%
      dplyr::select(year, ctr, iso3, everything()) %>%
      as.data.frame()
  
  } else {
    cat('\t TBD... and skip ... \n\n')
  }
  
  
  net_in <- func_net_import_direction(df = df_scenario, in_or_out = in_out, direction = drt, frac = 0.025)
  net_imports <- rbind(net_imports, cbind(ind = ind, net_in))
  
}

net_imports$scenario <- scenario
net_imports_fromNer  <- net_imports
```

  
  
  
```{r - nearby only scenario --> distant impact, include=FALSE}

scenario <- 'fromDst'
matrix   <- dst.matrix

### put all spillover indicators in one table 
net_imports <- data.frame()


### this loop process is the same as "nearby only" scenario above
for (i in 1:nrow(direction_use)) {
  # print(i)
  ind <- direction_use[i, 'ind']       %>% unlist() %>% paste(sep = '', collapse = ''); 
  drt <- direction_use[i, 'direction'] %>% unlist() %>% paste(sep = '', collapse = '') %>% as.numeric()
  in_out <- direction_use[i, 'Impact_direction'] %>% unlist() %>% paste(sep = '', collapse = '')
  print(ind)
  print(drt)
  
  
  ### if the direction is specified, read data; if not, skip. 
  if (!is.na(drt)) {
    xls <- paste0(dir.eora_cleaned, ind, '.xlsx'); print(xls)
    df  <- readxl::read_excel(path = xls)  %>% 
      gather(key = 'd', value = 'value', 4:ncol(.)) %>%
      as.data.frame()
    
    df_wide <- df %>%
      dplyr::group_by(year) %>%
      arrange(year, iso3, d) %>%
      spread(key = 'd', value = 'value') %>%
      arrange(year, iso3) %>%
      as.data.frame()
      
    ### --> `df` under different scenarios
    mat <- matrix %>%
      rownames_to_column(var = 'o') %>%
      gather(key = 'd', value = 'scenario', 2:ncol(.))
    
    df_scenario <- merge(df, mat, by.x = c('iso3', 'd'), by.y = c('o', 'd'), all.x = T) %>%
      dplyr::mutate(value = value * scenario) %>%
      dplyr::select(-scenario) %>%
      spread(key = 'd', value = 'value') %>%
      arrange(year, iso3) %>%
      dplyr::select(year, ctr, iso3, everything()) %>%
      as.data.frame()
  } else {
    cat('\t TBD... and skip ... \n\n')
  }    
  
  net_in <- func_net_import_direction(df = df_scenario, in_or_out = in_out, direction = drt, frac = 0.025)
  net_imports <- rbind(net_imports, cbind(ind = ind, net_in))
  
}



net_imports$scenario <- scenario
net_imports_fromDst  <- net_imports
``` 







#### - Save R data

```{r}

###' unify the year list ------------------------------------------------------------------
###' Use the number of years in EORA to determine how many years of data we need to include 
###' in this analysis 
f <- paste0(dir.flowScenario, 'para_n_year_check.RData'); f
load(f) ## `n_year_check` -- how many years of data are included in EORA
n_year_check

if (n_year_check == 4) {
  year_list <- c(2000, 2005, 2010, 2015)
  } else {
  year_list <- c(2015, 2019)
  }
year_list


net_imports_fromDst_fromNer <- rbind(net_imports_fromNer, net_imports_fromDst) %>%
  dplyr::filter(year %in% year_list) %>%
  ### use SHDB 2019 for 2015 --------------------!!!
  dplyr::mutate(
    year = ifelse(str_detect(string = ind, pattern = 'SHDB') & year == 2019, 
                  2015, year)) %>%
  as.data.frame()


## count how many indicators are included 
# net_imports_ind <- net_imports_fromDst_fromNer %>%
#   ungroup() %>%
#   dplyr::distinct(ind, .keep_all = F) %>%
#   dplyr::mutate(ind = trimws(ind)) %>%
#   arrange(ind)

## shdb do not have data in 2000, 2005, 2010
net_imports_bind_check <- net_imports_fromDst_fromNer %>%
  dplyr::filter(!year %in% c(2000, 2005, 2010, 2015)) %>%
  dplyr::distinct(ind, year, .keep_all = T)

### save the data
n_ft_ind <- length(unique(net_imports_fromDst_fromNer$ind));  n_ft_ind
fname <- paste0(dir.flowScenario, 'net_imports_fromDst_fromNer_', today, '_', n_ft_ind, '.RData'); fname
save(net_imports_fromDst_fromNer, file = fname)
```
