==> Code/22_22_FootprintDistance.Rmd <==
---
title: "Untitled"
author: "Yingjie"
date: "8/27/2021"
output: html_document
editor_options: 
  chunk_output_type: inline
---


# Setup
```{r include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

### To clear your environment 
remove(list = ls())

### data path
source('./Code/_path of data.R')
source('./Code/_package list.R')
source('https://raw.githubusercontent.com/Yingjie4Science/functionX/main/scaling_normalization_Pro.R')

library(dplyr)
library(ggplot2)
library(cowplot)

today <- format(Sys.time(), "%Y%m%d"); today
```




```{r}
### plot settings
unit_ns    <- 'cm'
width_1col <- 8.8   ## 1-column
width_2col <- 18    ## 2-column
font       <- 'sans'     ## "TT Arial"
font_size  <- 8          ##  Nature Sustainability: max = 7; min = 5
theme_ns <- 
  theme_bw()+
  theme(
    panel.grid.minor = element_blank(),
    legend.background = element_rect(fill="transparent"),
    legend.key.height = unit(0.15,"cm"),
    legend.key.width  = unit(0.15,"cm"),
    text = element_text(size=font_size + 2)
        )
```


# Data input


## SI data

```{r - income group}

load('./Data/Ancillary_Data_ISO3code_shp.RData') ## iso_eora, pop, gdp, shp, grp, grp_update

grp_use <- left_join(
  x = grp_update, 
  y = grp %>%
    dplyr::select(name, iso_a3, inc_group),
  by = c('name', 'iso_a3')
)

grp_id <- grp_use %>%
  
  ### re-code income group into 2 categories
  dplyr::mutate(group_numID = case_when(
    str_detect(group_income4, "High")         ~ "1",
    str_detect(group_income4, "Upper")        ~ "2",
    str_detect(group_income4, "Lower middle") ~ "3",
    str_detect(group_income4, "Low income")   ~ "4",
    TRUE                                      ~ '')) %>%
  
  dplyr::mutate(group_id = as.numeric(gsub("\\D", "", group_numID))) %>% 
  dplyr::select(iso_a3, group_id) %>%
  dplyr::filter(!is.na(iso_a3)) %>%
  as.data.frame()

# 1: high 
# 2: upper middle
# 3: low middle
# 4: low
```



## CEPII distance data
  
  https://pacha.dev/cepiigeodist/reference/dist_cepii.html
  
  dist        Simple distance (most populated cities, km)

  distcap     Simple distance between capitals (capitals, km)
  
  distw       Weighted distance (pop-wt, km) with theta=1 (theta measures the sensitivity of trade flows to bilateral distance dkl)
  
  distwces    Weighted distance (pop-wt, km) theta=-1
  
```{r - prep data}

library(cepiigeodist)
### filter countries that share borders
# dist_cepii[dist_cepii$contig == 1, ]

dist_cepii <- dist_cepii %>%
  dplyr::mutate(iso_o = ifelse(iso_o == 'ZAR', 'COD', iso_o), iso_d = ifelse(iso_d == 'ZAR', 'COD', iso_d))

geo_cepii  <- geo_cepii


### corrections to the `geo` data --------------------------------------------------------
s <- 'Congo'          ## --> code change
s <- 'Romania'        ## --> code change
s <- 'Serbia'         ## --> code change
s <- 'Monaco'         ## not included in this data, use 'CHE' as a proxy (nearby and similar size)
s <- 'Liechtenstein'  ## not included in this data, use 'CHE' as a proxy (nearby and similar size)
s <- 'Cura'           ## --> to add
s <- 'Montenegro'     ## --> to add
s <- 'Sudan'          ## --> to add 'South Sudan', gained independence from the Republic of the Sudan in 2011
s <- 'Gaza'
s <- 'Palestine'      ## --> code change
geo_cepii %>%
  dplyr::filter(str_detect(pattern = s, country))


geo_cuw <- geo_cepii %>% 
  dplyr::filter(iso3== 'ANT') %>% 
  head(1) %>%
  dplyr::mutate(iso2 = ifelse(iso2 == 'AN', 'CU', iso2), 
                iso3 = ifelse(iso3 == 'ANT', 'CUW', iso3))

dist_cuw <- dist_cepii %>% 
  dplyr::filter(iso_o == 'ANT'| iso_d == 'ANT') %>% 
  dplyr::mutate(iso_o = ifelse(iso_o == 'ANT', 'CUW', iso_o), iso_d = ifelse(iso_d == 'ANT', 'CUW', iso_d)) %>%
  as.data.frame()

geo_mne <- geo_cepii %>% 
  dplyr::filter(iso3== 'YUG') %>% 
  head(1) %>%
  dplyr::mutate(iso2 = ifelse(iso2 == 'YU', 'ME', iso2), 
                iso3 = ifelse(iso3 == 'YUG', 'MNE', iso3))
dist_mne <- dist_cepii %>% 
  dplyr::filter(iso_o == 'YUG'| iso_d == 'YUG') %>% 
  dplyr::mutate(iso_o = ifelse(iso_o == 'YUG', 'MNE', iso_o), iso_d = ifelse(iso_d == 'YUG', 'MNE', iso_d)) %>%
  as.data.frame()

geo_ssd <- geo_cepii %>% 
  dplyr::filter(iso3== 'SDN') %>% 
  head(1) %>%
  dplyr::mutate(iso2 = ifelse(iso2 == 'SD', 'SS', iso2), iso3 = ifelse(iso3 == 'SDN', 'SSD', iso3))
dist_ssd <- dist_cepii %>% 
  dplyr::filter(iso_o == 'SDN'| iso_d == 'SDN') %>% 
  dplyr::mutate(iso_o = ifelse(iso_o == 'SDN', 'SSD', iso_o), iso_d = ifelse(iso_d == 'SDN', 'SSD', iso_d)) %>%
  as.data.frame()

geo_sun <- geo_cepii %>% 
  dplyr::filter(iso3== 'RUS') %>% 
  head(1) %>%
  dplyr::mutate(iso2 = ifelse(iso2 == 'RU', 'SU', iso2), iso3 = ifelse(iso3 == 'RUS', 'SUN', iso3))
dist_sun <- dist_cepii %>% 
  dplyr::filter(iso_o == 'RUS'| iso_d == 'RUS') %>% 
  dplyr::mutate(iso_o = ifelse(iso_o == 'RUS', 'SUN', iso_o), iso_d = ifelse(iso_d == 'RUS', 'SUN', iso_d)) %>%
  as.data.frame()


geo_lie <- geo_cepii %>% 
  dplyr::filter(iso3== 'CHE') %>% 
  head(1) %>%
  dplyr::mutate(iso2 = ifelse(iso2 == 'CH', 'LI', iso2), iso3 = ifelse(iso3 == 'CHE', 'LIE', iso3))
dist_lie <- dist_cepii %>% 
  dplyr::filter(iso_o == 'CHE'| iso_d == 'CHE') %>% 
  dplyr::mutate(iso_o = ifelse(iso_o == 'CHE', 'LIE', iso_o), iso_d = ifelse(iso_d == 'CHE', 'LIE', iso_d)) %>%
  as.data.frame()


geo_mco <- geo_cepii %>% 
  dplyr::filter(iso3== 'CHE') %>% 
  head(1) %>%
  dplyr::mutate(iso2 = ifelse(iso2 == 'CH', 'MC', iso2), iso3 = ifelse(iso3 == 'CHE', 'MCO', iso3))
dist_mco <- dist_cepii %>% 
  dplyr::filter(iso_o == 'CHE'| iso_d == 'CHE') %>% 
  dplyr::mutate(iso_o = ifelse(iso_o == 'CHE', 'MCO', iso_o), iso_d = ifelse(iso_d == 'CHE', 'MCO', iso_d)) %>%
  as.data.frame()



### ---
geo  <- geo_cepii %>%
  dplyr::mutate(iso3 = ifelse(iso3 == 'ZAR', 'COD', iso3), iso2 = ifelse(iso2 == 'ZR', 'CD', iso2)) %>%
  dplyr::mutate(iso3 = ifelse(iso3 == 'ROM', 'ROU', iso3)) %>%
  dplyr::mutate(iso3 = ifelse(iso3 == 'YUG', 'SRB', iso3), iso2 = ifelse(iso2 == 'YU', 'RS', iso2)) %>%
  dplyr::mutate(iso3 = ifelse(iso3 == 'PAL', 'PSE', iso3)) %>%
  rbind(., geo_cuw, geo_mne, geo_ssd, geo_sun, geo_lie, geo_mco)



### corrections to the `dist` data -------------------------------------------------------

dist <- dist_cepii %>%
  dplyr::mutate(iso_o = ifelse(iso_o == 'ZAR', 'COD', iso_o), iso_d = ifelse(iso_d == 'ZAR', 'COD', iso_d)) %>%
  dplyr::mutate(iso_o = ifelse(iso_o == 'ROM', 'ROU', iso_o), iso_d = ifelse(iso_d == 'ROM', 'ROU', iso_d)) %>%
  dplyr::mutate(iso_o = ifelse(iso_o == 'YUG', 'SRB', iso_o), iso_d = ifelse(iso_d == 'YUG', 'SRB', iso_d)) %>%
  dplyr::mutate(iso_o = ifelse(iso_o == 'PAL', 'PSE', iso_o), iso_d = ifelse(iso_d == 'PAL', 'PSE', iso_d)) %>%
  rbind(., dist_cuw, dist_mne, dist_ssd, dist_sun, dist_lie, dist_mco) %>%
  dplyr::distinct(iso_o, iso_d, .keep_all = T)



head(dist)

# d <- dist$dist %>% as.numeric() ## delete?

ctr_list_cepii <- dist %>%
  dplyr::select(1:2) %>%
  gather(key = 'od', value = 'iso', 1:2) %>%
  dplyr::distinct(iso, .keep_all = T) %>%
  arrange(iso)

```


```{r - distance stats}

## load data from `helper_adjancet_distant_matrix.Rmd`
f <- './Data/_neighbor_matrix/_input_ctr_matrix_neighbor.xlsx'
neighbor_matrix <- readxl::read_excel(path = f) 

## convert to od (long) format
neighbor_matrix_od <- neighbor_matrix %>%
  dplyr::select(-Country, -Neighbors) %>%
  pivot_longer(cols = 2:ncol(.), names_to = 'd', values_to = 'adjacent') %>%
  ## add distance data from cepii
  merge(x = ., y = dist %>% dplyr::select(-c('comlang_off':'smctry', 'distcap':'distwces')), 
        by.x = c('o', 'd'), by.y = c('iso_o', 'iso_d'), all.x = T) %>%
  ## to check the consistency between the two classifications
  dplyr::mutate(consistency = ifelse(adjacent == contig, 1, 0))

## count the number of consistency
neighbor_matrix_od %>% count(consistency)


## average distance of nearby (or distant) countries, unit = km
neighbor_matrix_od %>% 
  ## remove self-loop
  dplyr::filter(o != d) %>%
  summarise_at(vars(dist), list(mean = mean, median = median, min = min, max = max), na.rm=TRUE)

neighbor_matrix_od %>% 
  ## remove self-loop
  dplyr::filter(o != d) %>%
  group_by(adjacent) %>%
  # group_by(contig) %>%
  summarise_at(vars(dist), list(mean = mean, median = median, min = min, max = max), na.rm=TRUE)



neighbor_matrix_od_group <- neighbor_matrix_od %>%
  dplyr::mutate(dist_e3 = dist/10^3) %>%
  Rmisc::summarySE(measurevar= "dist_e3", groupvars = c('adjacent'), na.rm = T) %>%
  as.data.frame()

```




## Footprint data

```{r}
## load ft indicator list and direction data ----------------------
f <- paste0(dirname(dir.eora_cleaned), '/', 'direction_fts.rds'); f
direction_fts <- readRDS(file = f)

f <- paste0(dirname(dir.eora_cleaned), '/', 'direction.RData')
load(f) ## `direction`, `direction_clean`

### loop the normalization based on `direction` ------------------------------------------
ft_ind <- direction_clean %>%
  dplyr::select(ind, SDG_class, ShortName, direction, everything()) %>%
  dplyr::filter(!is.na(direction)) %>%
  dplyr::filter(Used == '1') %>%
  arrange(SDG_class)

cat('\n', length(unique(ft_ind$ind)), 'indicators were included in the analysis.\n')
```



```{r include=FALSE}
### !!! ** Choose how many type of flows will be used**
class_detail <- F   ## use aggregated 
# class_detail <- T

###' class-detailed | class-aggregated 
###' ---------------------------------- #
###' high on low    | Vertical
###' high on high   | Vertical
###' low on high    | Horizontal
###' low on low     | Horizontal
###' -----------------------------------#


### --> pick an ft indicator as a test -----
# ind <- 'Water'
# ind <- 'student_flow'
# ind <- 'Landuse_forest'
# ind <- 'osh_fatal'


links_dist_type_collection <- data.frame()

for (ind in unique(ft_ind$ind)) {
  
  ## test
  # ind <- "CO2"
  
  print(ind)
  
  xls <- paste0(dir.eora_cleaned, ind, '.xlsx'); xls
  
  ## if the data are from SHDB, we use 2019 to represent 2015
  is_shdb <- str_detect(string = ind, pattern = "SHDB")
  
  if (is_shdb == F) {
    mat <- readxl::read_excel(xls) %>% 
      ungroup() %>%
      dplyr::filter(year == 2015) 
  } else {
    mat <- readxl::read_excel(xls) %>% 
      ungroup() %>%
      dplyr::mutate(year = ifelse(year == 2019, 2015, year)) %>%
      dplyr::filter(year == 2015) 
  }
  
  ## to check if all the data have been loaded in 
  if (nrow(mat) < 1) {
    cat('\n\t This indicator has misisng data, pls check! \n')
  }
  
  links <- mat %>%
    gather(key = 'to', value = 'weight', 4:ncol(.)) %>%
    dplyr::rename('from' = 'iso3') %>%
    dplyr::mutate(type = ind) %>%
    dplyr::select(-ctr, -year) %>%
    arrange(from, to) %>%
    dplyr::distinct(from, to, .keep_all = T) %>%  
    dplyr::select(from, to, everything()) %>%
    as.data.frame()
  
  ##' need to normalize the weight factor for calculation
  ##' --> If iqr == 0, `robust_scaling` will not work as iqr will be used as the denominator
  ##'     in the function.
  ##' --> To address the issue, we will use Z-Score Scaling when iqr == 0.
  which_scaling <- IQR(links$weight, na.rm = TRUE)
  if (which_scaling == 0) {
    links_w <- func_z_scaling(data = links, var = 'weight') %>%
      dplyr::mutate(weight = func_norm_minmax(weight, a=0.01, b=100))
  } else {
    links_w <- func_robust_scaling(data = links, var = 'weight') %>%
      dplyr::mutate(weight = func_norm_minmax(weight, a=0.01, b=100))
  }
  
  links_dist <- 
    merge(x = links_w, y = dist,   
          by.x = c('from', 'to'), by.y = c('iso_o', 'iso_d'), all.x = T) %>%
    merge(x = ., y = grp_id, 
          by.x = 'from', by.y = 'iso_a3', all.x = T) %>% dplyr::rename('group_from' = 'group_id') %>%
    merge(x = ., y = grp_id, 
          by.x = 'to',   by.y = 'iso_a3', all.x = T) %>% dplyr::rename('group_to' = 'group_id') %>%
    
    ## check NA in iso3 matches
    arrange(!is.na(group_from), !is.na(group_to)) %>%
    
    dplyr::mutate(
      group_od = paste0(group_from, group_to), ## group_od = 14, means o-1 sending; d-4 receiving 
      group_od = as.numeric(group_od)
      ) %>%
    dplyr::select(-group_from, -group_to) %>%
    arrange(to, from)
  
  
  ### --> check data 
  # links_dist_na <- links_dist %>% dplyr::filter(is.na(dist))
  # 
  # ctr_match_check <- mat %>%
  #   dplyr::select(2:3) %>%
  #   distinct(iso3, .keep_all = T) %>%
  #   merge(., ctr_list_cepii, by.x = 'iso3', by.y = 'iso', all.x = T) %>%
  #   arrange(!is.na(od))
  
  ### --> if we want to see a detailed classification of one group's impact on the other
  links_dist_detail <- links_dist %>%
    dplyr::mutate(
      od_direction = case_when(
      group_od %in% c(14, 13, 12) ~ 'low on high', ## 'vertical flow'
      group_od %in% c(41, 31, 21) ~ 'hign on low', ## 'horizontal flow'
      TRUE ~ 'other') 
    ) %>% as.data.frame()
    
  
  ### --> if we want to only see if the interaction happen across large income difference groups
  links_dist_aggr <- links_dist %>%
    dplyr::mutate(
      od_direction = case_when(
        
        ## option 1
        # group_od %in% c(12, 13, 14) ~ 'Vertical flows', ## 
        # group_od %in% c(21, 31, 41) ~ 'Vertical flows', ## 
        
        ## option 2
        group_od %in% c(12, 13, 14, 23, 24, 34) ~ 'Vertical flows', ## 
        group_od %in% c(21, 31, 41, 32, 42, 43) ~ 'Vertical flows', ## 
      
      TRUE ~ 'Horizontal flows') 
    ) %>% as.data.frame()
  
  
  
  ### !!! choose one for analysis
  if (class_detail == F) {
    links_dist_type <- links_dist_aggr
  } else if (class_detail == T) {
    links_dist_type <- links_dist_detail
  } else {
    print('NA ...................')
    break
  }
  
  
  links_dist_type_collection <- rbind(links_dist_type_collection, links_dist_type)
  
}

cat(length(unique(links_dist_type_collection$type)), 'indicators were included in the analysis.\n')


## save result
f <- paste0(dir_results, 'links_dist_type_collection.rds') 
saveRDS(object = links_dist_type_collection, file = f)
```




# Analysis - FT distance 

  For country *m*, on average, how far has its impact reached to? 
  
  Here we use the *footprint distance* to measure the *impact distance*.
  
## distance measure  
```{r}

## load data
f <- paste0(dir_results, 'links_dist_type_collection.rds') 
links_dist_type_collection <- readRDS(file = f)

##' check on data to see the distribution, and if there are NA values, or negative values
##'     in the data. 
hist(links_dist_type_collection$weight, nclass = 20)

links_dist_type_collection %>%
  ggplot(., aes(weight)) +
  geom_histogram()

links_dist_type_collection %>%
  dplyr::filter(weight < 10^3) %>%
  dplyr::filter(weight > -10^3) %>%
  ggplot(., aes(weight)) +
  geom_histogram() +
  geom_vline(xintercept = 0, color = 'red')

links_dist_type_collection_check <- links_dist_type_collection %>%
  dplyr::filter(weight <0)

unique(links_dist_type_collection_check$type)
##' only "FDI_matrix" and "ODA_OOF_oecd_matrix" has negative values
##' to avoid errors, we set these negative values as NA (or remove these rows) when calculating
##'   impact distance because technically negative value mean the country 
##'   is impacted but not impacting others


## for a receiving country, calculate its impact distance --------------------------------- 
ft_d <- links_dist_type_collection %>%
  dplyr::filter(weight > 0) %>%                           ## remove negative values - see notes above
  dplyr::mutate(ftd = weight * dist) %>%                  ## flow * distance
  dplyr::group_by(to, type, od_direction) %>%
  dplyr::summarise(dq_sum = sum(ftd, na.rm = T),          ## total (flow * distance)
                   q_sum  = sum(weight, na.rm = T),       ## total flow
                   d_avg_geo = mean(dist, na.rm = T)
                   ) %>%   
  ungroup() %>%
  as.data.frame() %>%
  dplyr::mutate(d_avg_ft = dq_sum/q_sum) %>%              ## average distance (km) of each unit of flows
  merge(x = ., 
        y = ft_ind %>% dplyr::select(-direction), 
        by.x = 'type', by.y = 'ind', all.x = T)


links_dist_type_collection %>% 
  dplyr::group_by(type, od_direction) %>%
  tally()



od_direction_levels <- unique(links_dist_type_collection$od_direction); od_direction_levels
od_direction_count  <- length(od_direction_levels); od_direction_count
if (od_direction_count <= 2) {
  od_direction_levels <-  rev(od_direction_levels)
} else {
  od_direction_levels <-  c("hign on low", "low on high", "other")
}
od_direction_labels <- str_to_sentence(od_direction_levels)


## 1. mean distance between countries 
hist(dist_cepii$dist)
mean(dist_cepii$dist, na.rm=T) ## --> 8481.7 km

## 2. mean impact distance between countries 
hist(ft_d$d_avg_ft)
mean(ft_d$d_avg_ft, na.rm=T)   ## --> 4948.6 km


mean(ft_d$d_avg_geo, na.rm=T)  ## --> 5912.5 km


first_upper <- function(x) {
  substr(x, 1, 1) <- toupper(substr(x, 1, 1))
  x
}

ft_ind_ls <- unique(ft_ind$ind); ft_ind_ls
ft_ind_ls_levels <- ft_ind_ls
ft_ind_ls_labels <- ft_ind_ls_levels %>% 
  gsub('_|TradeMatrix|Annual crops|irri', ' ', x = .) %>%
  gsub('Osh', 'OSH ', x = .) %>%
  gsub('Energy', 'Energy ', x = .) %>%
  gsub('Landuse', 'Landuse ', x = .) %>%
  gsub('Nitrogen', 'Nitrogen ', x = .) %>%
  gsub('Phosphorus', 'Phosphorus ', x = .) %>%
  trimws(.) %>%
  first_upper(.)
```




```{r plot fig ex6 - by ft indicator}

### 1. use ft distance
###'   used this footprint distance before 01/01/2024
prefix <- 'ft'
ft_d_i <- ft_d %>%
  dplyr::rename('d_avg_select' = "d_avg_ft") %>%
  dplyr::mutate(d_avg_select = d_avg_select/10^3) %>%
  as.data.frame()


### 2. use geographic distance 
###'  Better to use geographic distance so that we can also add the mean distance of adjacent countries, 
###'    and the mean distance of distant countries. 
prefix <- 'geo'
ft_d_i <- ft_d %>%
  dplyr::rename('d_avg_select' = "d_avg_geo") %>%
  dplyr::mutate(d_avg_select = d_avg_select/10^3) %>%
  as.data.frame()


### plot ---------------------------------------------------------------------------------
d_max <- max(ft_d_i$d_avg_select, na.rm=T) %>% ceiling(.); d_max ## --> 20 
n_plots  <- length(unique(ft_d$type))
p_height <- ceiling(n_plots/4) * 7/8

p1 <- ft_d_i %>% 
  # dplyr::filter(SDG_class != 'Soci') %>% ## subset for easy plot
  dplyr::mutate(od_direction = factor(x = od_direction, levels = od_direction_levels, labels = od_direction_labels)) %>%
  
  ### - used this before 2023-07-01
  # dplyr::mutate(type = factor(x = type, levels = ft_ind_ls_levels, labels = ft_ind_ls_labels)) %>%
  ### - change on since  2023-07-17
  dplyr::mutate(type = ShortName) %>%
  arrange(type) %>%
  dplyr::mutate(type = as.factor(type)) %>%
  
  ### - plot
  ggplot() +
  geom_histogram(aes(x = d_avg_select, fill = od_direction), position="identity", alpha=0.5, show.legend = T) +
  geom_vline(aes(xintercept=mean(d_avg_select, na.rm=T)), color="red", linewidth = .5) + # linetype="dashed", alpha = .5,
  
  # ggplot2::annotate(
  #   "text", 
  #   x = round(mean(ft_d_i$d_avg_select, na.rm=T)*1, digits = 1),
  #   y = 2, 
  #   label = round(mean(ft_d_i$d_avg_select, na.rm=T), digits = 1), 
  #   colour = "red", size = 2) +
  # facet_wrap(~od_direction, ncol = 1) + ## for one indicator
  # facet_grid(type~od_direction) +
  facet_wrap(~type, ncol = 4, scales = 'free_y', labeller = labeller(type = label_wrap_gen(31))) +
  theme_ns +
  # theme_minimal() +
  theme(
    # axis.text.x = element_text(angle = 45, hjust = 1),
    legend.position = c(0.99, 0.01),
    legend.justification = c(1, 0), # right bottom justification
    legend.box.margin = margin(r = 0, b = 0, unit = "mm"), # small margin
    strip.text = element_text(size = 8),
    axis.title.y = element_blank(), 
    panel.grid = element_blank(),
    ) +
  scale_y_continuous(n.breaks = 4) +
  scale_x_continuous(
    breaks = c(seq(0, d_max+.5, 9), round(mean(ft_d_i$d_avg_select, na.rm=T), digits = 1))) +
  xlab(label = bquote('Distance (10'^3~'km)')) + 
  labs(fill='') # Impact pairs


ggarrange(p1, labels = '')
fname <- paste0(dir.fig, '01main/', 'ex_6_ImpactDistance_byFT',
                '_class', od_direction_count, '_', today, '_', prefix, '.jpg'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = 7, height = p_height, units = 'in', dpi = 300)
```






```{r plot fig 2d - summary across indicators}
ind <- 'AvgAcrossFT'

### 1. use ft distance
###'   used this footprint distance before 01/01/2024
# prefix <- 'ft'
# ft_d_summary <- ft_d %>%
#   dplyr::rename('d_avg_select' = "d_avg_ft") %>%
#   Rmisc::summarySE(measurevar= "d_avg_select", groupvars = c('to', 'od_direction'), na.rm = T) %>%
#   as.data.frame()
# 
# ft_d_summary_group <- ft_d %>%
#   dplyr::rename('d_avg_select' = "d_avg_ft") %>%
#   Rmisc::summarySE(measurevar= "d_avg_select", groupvars = c('od_direction'), na.rm = T) %>%
#   as.data.frame() %>%
#   dplyr::mutate(od_direction = factor(x = od_direction, levels = od_direction_levels, labels = od_direction_labels)) %>%
#   data.table::as.data.table()
# 
# ft_d_i <- ft_d %>% 
#   dplyr::rename('d_avg_select' = "d_avg_ft") %>%
#   dplyr::mutate(d_avg_select = d_avg_select/10^3) %>%
#   as.data.frame()



### 2. use geographic distance 
###'  Better to use geographic distance so that we can also add the mean distance of adjacent countries, 
###'    and the mean distance of distant countries. 
prefix <- 'geo'
ft_d_summary <- ft_d %>%
  dplyr::rename('d_avg_select' = "d_avg_geo") %>%
  dplyr::mutate(d_avg_select = d_avg_select/10^3) %>%
  Rmisc::summarySE(measurevar= "d_avg_select", groupvars = c('to', 'od_direction'), na.rm = T) %>%
  as.data.frame()

ft_d_summary_group <- ft_d %>%
  dplyr::rename('d_avg_select' = "d_avg_geo") %>%
  dplyr::mutate(d_avg_select = d_avg_select/10^3) %>%
  Rmisc::summarySE(measurevar= "d_avg_select", groupvars = c('od_direction'), na.rm = T) %>%
  # as.data.frame() %>%
  dplyr::mutate(od_direction = factor(x = od_direction, levels = od_direction_levels, labels = od_direction_labels)) %>%
  data.table::as.data.table()

ft_d_i <- ft_d %>%
  dplyr::rename('d_avg_select' = "d_avg_geo") %>%
  dplyr::mutate(d_avg_select = d_avg_select/10^3) %>%
  as.data.frame()



### plot -------------------------------------- 
d_max_ <- max(ft_d_summary$d_avg_select, na.rm=T) %>% ceiling(.); d_max_  ## --> 15

ft_d_summary %>% 
  group_by(od_direction) %>%
  summarise_at(vars(d_avg_select), list(mean = mean, median = median, min = min, max = max), na.rm=TRUE)



## mean distance for the Horizontal and Vertical groups
mean_d_h <- ft_d_summary_group[od_direction == 'Horizontal flows', d_avg_select]
mean_d_v <- ft_d_summary_group[od_direction == 'Vertical flows', d_avg_select]


neighbor_matrix_od_group_help <- neighbor_matrix_od_group %>%
  dplyr::mutate(od_direction = ifelse(adjacent == 0, 'Vertical flows', 'Horizontal flows')) %>%
  dplyr::mutate(od_direction = factor(x = od_direction, levels = od_direction_levels, labels = od_direction_labels)) %>%
  data.table::as.data.table()

p <- ft_d_summary %>% 
  # dplyr::filter(SDG_class != 'Soci') %>% ## subset for easy plot
  dplyr::mutate(od_direction = factor(x = od_direction, levels = od_direction_levels, labels = od_direction_labels)) %>%
  ggplot() +
  geom_histogram(aes(x = d_avg_select, fill = od_direction), position="identity", alpha=0.4, show.legend = F) +
  
  ## add the average distance among all countries
  geom_vline(aes(xintercept= round(mean(ft_d_i$d_avg_select, na.rm=T), digits = 1) ), 
             color="red", alpha = .7, size = 0.5) + # linetype="dashed", 
  
  ## add the average distance of Horizontal and Vertical groups
  ### 1. use mean ft distance by v and h
  # geom_vline(data = ft_d_summary_group, aes(xintercept=d_avg_select, color = od_direction), linetype="dashed", show.legend = F) +
  ### 2. use mean ft distance by v and h + SD error bar
  geom_point(data = ft_d_summary_group, aes(x=d_avg_select, y = 1, color = od_direction), show.legend = F) +
  geom_errorbarh(data = ft_d_summary_group, 
                 aes(xmax = d_avg_select + sd, xmin = d_avg_select - sd, y = 1, color = od_direction), 
                 height = 1.5, alpha = .8, linewidth = 0.4, show.legend = F) +
  ### 3. use mean geographic distance by nearby and distant
  # geom_point(data = neighbor_matrix_od_group_help, aes(x=dist_e3, y = 5, color = od_direction), show.legend = F) +
  # geom_errorbarh(data = neighbor_matrix_od_group_help, 
  #                aes(xmax = dist_e3 + sd, xmin = dist_e3 - sd, y = 5, color = od_direction), 
  #                height = 2, alpha = .5, show.legend = F) +

  facet_wrap(~od_direction, ncol = 1) + ## for one indicator
  theme_ns +
  # theme_minimal() +
  theme(
    # axis.text.x = element_text(angle = 45, hjust = 1),
    axis.title.y = element_blank(), 
    panel.grid = element_blank(),
    ) +
  # ylim(0,20) + 
  scale_x_continuous(
    # limits = c(0, d_max_), 
    breaks = c(seq(0, d_max_, 3.5), round(mean(ft_d_i$d_avg_select, na.rm=T), digits = 1))) +
  xlab(label = bquote('Distance (10'^3~'km)')) + 
  labs(fill='Impact pair')
ggarrange(p, labels = 'd')
fname <- paste0(dir.fig, '01main/', 'fig2d_ImpactDistance_', ind, '_',
                od_direction_count, 'grp_', today, '_', prefix, '.jpg'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = 7.5/2, height = 7.5/2, units = 'in', dpi = 300)
```



### count % of nearby (contiguous) or not
```{r}

## impact distance by group of 'nearby or not'
ft_FarNear <- links_dist_type_collection %>%
  dplyr::group_by(contig) %>%
  dplyr::summarise(q_sum  = sum(weight, na.rm = T)) %>%
  as.data.frame() %>%
  dplyr::mutate(
    sum = sum(links_dist_type_collection$weight, na.rm = T),
    pct = q_sum/sum
    ) %>%
  as.data.frame()


## number of countries that are nearby
ft_FarNear_n <- links_dist_type_collection %>%
  dplyr::group_by(contig) %>%
  dplyr::summarise(n = n()) %>%
  as.data.frame() %>%
  dplyr::mutate(
    sum = sum(n, na.rm = T),
    pct = n/sum
    ) %>%
  as.data.frame()

```
