==> Code/26_TNI_on_SDGs_Viz.Rmd <==
---
title: "Spillover on SDGs"
author: "Yingjie Li"
date: "2021-03-01 (Updated: 2021-03-04)"
output: pdf_document
editor_options: 
  chunk_output_type: inline
---


# Set up
```{r Dirs and packages, include=FALSE}

### To clear your environment 
remove(list = ls())

### set work dir
path <- rstudioapi::getSourceEditorContext()$path
dir  <- dirname(rstudioapi::getSourceEditorContext()$path); dir
setwd(dir)
setwd('..') # set directory by one folder up
getwd()


### data path
source('./Code/_path of data.R')
dir.input1 <- dir.cleaned       ## data for country profile, which are cleaned 
dir.input2 <- dir.flowScenario  ## data for no_trade, no_dst_trade, no_nearby_trade scenarios
dir.output <- './Data/data_03_results/'



### packages
source('./Code/_package list.R')

library(Rmisc) ## for `summarySE`
library(scales)
library(tmap)
library(RColorBrewer)
library(ggpattern)

### Disable Scientific Notation in R
options(scipen = 999) # Modify global options in R

today <- format(Sys.time(), "%Y%m%d"); today
```



```{r Theme and font}
theme_set(theme_bw())

### plot settings
unit_ns    <- 'cm'
width_1col <- 8.8   ## 1-column
width_2col <- 18    ## 2-column
font       <- 'sans'     ## "TT Arial"
font_size  <- 8          ##  Nature Sustainability: max = 7; min = 5
theme_ns <- 
  theme_bw()+
  theme(
    # axis.title =element_blank(),
    # axis.text  =element_blank(),
    # axis.ticks =element_blank(),
    # panel.background = element_rect(fill = NA),
    # panel.grid.major.x = element_blank(),
    # panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    legend.background = element_rect(fill="transparent"),
    legend.key.size = unit(0.15,"cm"),
    text = element_text(size=font_size)
        )

theme_map <- theme(
    axis.title = element_blank(),
    axis.text  = element_blank(),
    axis.ticks = element_blank(),
    panel.background = element_rect(fill = NA),
    panel.grid  = element_blank())


theme_bar_my <- theme(
  # axis.title = element_text(size = font_size+2),
  axis.text  = element_text(size = font_size+3))
```



```{r Colors}
### create map colors for score change
colors_changes <- c(
  ## brown
  '#f44343', 
  '#f46d43',
  '#feb24c',
  '#ffeda0',
  
  ## blues
  '#eff3ff', '#c6dbef', '#6baed6', '#4292c6' ## option1
  # '#e0f3f8', '#abd9e9', '#74add1', '#4575b4' ## option2

)
```




```{r Functions}
### normalization using the function
### Rather than just dropping the top and bottom trim percent, these extreme values are replaced with values at the trim and 1- trim quantiles. See more at https://rdrr.io/cran/psych/man/winsor.html
### We added one more step of this data processing than Xu and Li et al 2020 in NS. 
source("./Code/function_norm_sdg_score_good_bad.R")

```



## Ancillary data
```{r}
load('./Data/Ancillary_Data_ISO3code_shp.RData') ## `iso_eora`, `pop`, `gdp` (USD), `shp`, `grp`, `grp_update`
pop_l <- pop %>% as.data.frame() %>%
  gather(key = 'year', value = 'pop', `2000`:ncol(.)) %>%
  dplyr::select(-ctr) %>%
  as.data.frame()
```



# Data & Analysis

## SDG Score & Index

### Score & change
```{r - load data - choose ❓}
### load data
ls_f <- list.files(path = './Data/data_03_results', pattern = '^SDGs_all', full.names = T); ls_f
f <- tail(ls_f, n = 1); cat('\n Here, we load this version as input: \n\t', f, '\n\n')
load(f) 

unique(dfs$scenario)
snr_levels <- c("value_rel", "value_dst", "value_ner", "value_not")
# snr_labels <- c("real", "no nearby", "no distant", "none")
snr_labels <- c("Real world", "Telecoupling only", "Pericoupling only", "Intracoupling only") %>%
  str_to_sentence()
snr_colors <- c('#418463', '#5eab94', '#96c56f', '#dfe360')

dfs_score <- dfs %>%
  ungroup() %>%
  dplyr::select(-val) %>%
  dplyr::filter(year == 2015) %>%
  dplyr::filter(!iso3 %in% c('SUN', 'ROW')) %>%
  group_by(iso3, year, scenario, goal) %>%
  dplyr::summarise(score = mean(value_norm, na.rm = TRUE)) %>% ungroup() %>%
  dplyr::mutate(score    = as.numeric(score)) %>%
  dplyr::mutate(goal     = factor(x = goal, levels = seq(1, 17, 1)),
                scenario = factor(x = scenario, levels = snr_levels)) %>%
  ### add country group info (updated on 2023-01-17)
  merge(x = .,
        y = grp_update %>% 
          dplyr::rename(iso3=iso_a3) %>% 
          dplyr::select(iso3, group_income2) %>%
          dplyr::rename(group = group_income2) %>% ## based on MG's suggestion
          dplyr::distinct(iso3, .keep_all = T),
        by = 'iso3', all.x = T) %>%
  as.data.frame()
str(dfs_score)
cat('The final data include', length(unique(dfs_score$iso3)), 'countries.') ## 189


### 
f <- paste0(dir.output, "SDG_score_byGoal_byISO3_byYear.rds")
saveRDS(dfs_score, file = f)
```



```{r - score change & outlier}

### for each country, to what extent each goal was impacted by metacouplings? ------------
dfs_score_chg <- dfs_score %>%
  # dplyr::select(-group) %>%
  spread(key = scenario, value = score) %>%
  dplyr::mutate(
    value_chg = value_rel - value_not, 
    value_chg = ifelse(iso3 =='GRL', NA, value_chg)) %>%    ## Greenland
  dplyr::filter(year == 2015) %>%
  as.data.frame()


### outliers? ----------------------------------------------------------------------------
# Set up the plotting layout
par(mfrow = c(1, 2))  # 1 row, 2 columns

hist(dfs_score_chg$value_chg, nclass = 100)

library(remotes)
# Source the function from GitHub
url_link <- "https://raw.githubusercontent.com/Yingjie4Science/functionX/main/func_find_outliers_bounds.R"
source(url_link)
# find_outlier_bounds_Rosner(x = dfs_score_chg$value_chg)
find_outlier_bounds_Hampel(x = dfs_score_chg$value_chg)
find_outlier_bounds_iqr(x = dfs_score_chg$value_chg)
# lower     upper 
# -15.69421  24.36477 


### deal with these outliers by reducing the magnitude -----------------------------------
outlier_bounds<- find_outlier_bounds_iqr(x = dfs_score_chg$value_chg)
outlier_lower <- outlier_bounds[1]
outlier_upper <- outlier_bounds[2]

dfs_score_chg <- dfs_score_chg %>%
  dplyr::mutate(
    value_chg0 = value_chg,
    value_chg = ifelse(value_chg > outlier_upper,  
                       value_chg - (value_chg-outlier_upper)*0.75, 
                       value_chg),
    value_chg = ifelse(value_chg < outlier_lower, 
                       value_chg - (value_chg-outlier_lower)*0.75, 
                       value_chg),
    ) %>% 
  dplyr::mutate(value_chg_pct = value_chg/value_rel * 100) %>%    ## % of change
  # dplyr::select(-value_not, -value_rel) %>%
  as.data.frame()

hist(dfs_score_chg$value_chg, nclass = 100)
### check Goal id
vars <- unique(dfs_score_chg$goal) %>% as.numeric() %>% sort() %>% as.character(); vars

# Reset the plotting layout to the default
par(mfrow = c(1, 1))
```


```{r - % change by Goal - MS ✏️📝}

### deal with these outliers by reducing the magnitude -----------------------------------
outlier_bounds<- find_outlier_bounds_iqr(x = dfs_score_chg$value_chg_pct)
outlier_lower <- outlier_bounds[1]; outlier_lower
outlier_upper <- outlier_bounds[2]; outlier_upper
dfs_score_chg <- dfs_score_chg %>%
  dplyr::mutate(value_chg_pct = winsor(value_chg_pct, trim = 0.01, na.rm = TRUE)) %>% 
  dplyr::mutate(
    value_chg_pct = ifelse(value_chg_pct > outlier_upper,
                       value_chg_pct - (value_chg_pct-outlier_upper)*0.75,
                       value_chg_pct),
    value_chg_pct = ifelse(value_chg_pct < outlier_lower,
                       value_chg_pct - (value_chg_pct-outlier_lower)*0.75,
                       value_chg_pct)
    ) %>%
  as.data.frame()
hist(dfs_score_chg$value_chg_pct, nclass = 100)


chg_pct_err  <- dfs_score_chg %>% 
  dplyr::filter_if(~is.numeric(.), all_vars(!is.infinite(.))) %>%
  dplyr::filter_if(~is.numeric(.), all_vars(!is.na(.))) %>%
  Rmisc::summarySE(measurevar= "value_chg_pct", groupvars = c("year", "goal"), na.rm = T) %>%
  as.data.frame()

chg_pct_err %>%
  Rmisc::summarySE(measurevar= "value_chg_pct", groupvars = c("year"), na.rm = T)
```


```{r ___ fig 1b - bar - change by Goal}

### Bar plot - Goal score change - by goal ---------------------------------------------------------
chg_err  <- dfs_score_chg %>% 
  # dplyr::filter(group_2 != 'na') %>%
  Rmisc::summarySE(measurevar= "value_chg", groupvars = c("year", "goal"), na.rm = T)
# mycol  <-  scale_fill_manual(values = mycols)


### change by intervals for colors ------------
# library(classInt)
# brks <- c(-Inf, -20, -10, -5, 0, 5, 10, 20, Inf)
# chg_err$breaks <- cut(chg_err$value_chg, breaks=brks) #, labels=labs)
# levels(chg_err$breaks)
# unique(chg_err$breaks)
# mycols <- colors_changes[4:8]
levels(chg_err$value_chg)

p1b <- chg_err %>%
  dplyr::mutate(chg_direction = ifelse(value_chg > 0, 1, -1),
                chg_direction = factor(chg_direction, levels = c(1,-1))) %>%
  ggplot(data = ., 
         aes(y = factor(goal, levels = rev(vars)), x = value_chg, fill = chg_direction)) +
  geom_bar(position = position_dodge(0.8), stat="identity", width = 0.8, show.legend = F, alpha = .7)+
  geom_errorbar(aes(
    xmin=value_chg-se,
    xmax=value_chg+se), 
    width=.3, color = 'gray20',
    position=position_dodge(.8)) +
  # scale_fill_manual(values = mycols) +
  scale_fill_manual(values = c(#'black', 
                               '#6baed6',
                               '#ffeda0')) +
  scale_x_continuous(labels = scales::number_format(accuracy = 1)) +
  xlab("Transnational impact on SDG score")+ ylab('SDGs') + labs(fill = "Goal")+
  theme_bar_my +
  theme(plot.margin = unit(c(0.1, 0, 0.1, -.5), "cm"))+ 
  theme(legend.position = c(0.1, 0.85), 
        axis.ticks.y = element_blank(),
        # panel.grid.major = , 
        # axis.text.y = element_blank(),
        panel.grid.minor = element_blank())
p1b
# fname <- paste0(dir.fig, 'SDG_score_chg_errorBar.jpg'); fname
# ggsave(filename = fname, plot = last_plot(), width = 3, height = 5, units = 'in', dpi = 300)




### SDG icons ----------------------------------------------------------------------------

### read in SDG icons 
getwd()
image_png <- sort(sample(dir("../SDG_icons/UN/SDG-Icons-2019_WEB", 
                            full.names = TRUE), 17))
image_png
img_sur <- "../SDG_icons/UN/SDG-Icons-2019_WEB/E-WEB-Goal-"


### make the image strip ----------------- 
### SDG icon order: top->down: 17->1
sdg_id   <- seq(1,17,1); sdg_id
sdg_id.y <- sdg_id - 0.5; sdg_id.y  ## - 0.5
### reverse the order: top->down: 1->17
sdg_id.y <- rev(sdg_id.y); sdg_id.y
scale_here <- 0.9

icons <- axis_canvas(plot = p1b, axis = 'y') + 
  draw_image(paste0(img_sur, '01', '.png'), y = sdg_id.y[1], scale = scale_here) +
  draw_image(paste0(img_sur, '02', '.png'), y = sdg_id.y[2], scale = scale_here) +
  draw_image(paste0(img_sur, '03', '.png'), y = sdg_id.y[3], scale = scale_here) +
  draw_image(paste0(img_sur, '04', '.png'), y = sdg_id.y[4], scale = scale_here) +
  draw_image(paste0(img_sur, '05', '.png'), y = sdg_id.y[5], scale = scale_here) +
  draw_image(paste0(img_sur, '06', '.png'), y = sdg_id.y[6], scale = scale_here) +
  draw_image(paste0(img_sur, '07', '.png'), y = sdg_id.y[7], scale = scale_here) +
  draw_image(paste0(img_sur, '08', '.png'), y = sdg_id.y[8], scale = scale_here) +
  draw_image(paste0(img_sur, '09', '.png'), y = sdg_id.y[9], scale = scale_here) +
  draw_image(paste0(img_sur, 10, '.png'),   y = sdg_id.y[10], scale = scale_here) +
  draw_image(paste0(img_sur, 11, '.png'),   y = sdg_id.y[11], scale = scale_here) +
  draw_image(paste0(img_sur, 12, '.png'),   y = sdg_id.y[12], scale = scale_here) +
  draw_image(paste0(img_sur, 13, '.png'),   y = sdg_id.y[13], scale = scale_here) +
  draw_image(paste0(img_sur, 14, '.png'),   y = sdg_id.y[14], scale = scale_here) +
  draw_image(paste0(img_sur, 15, '.png'),   y = sdg_id.y[15], scale = scale_here) +
  draw_image(paste0(img_sur, 16, '.png'),   y = sdg_id.y[16], scale = scale_here) +
  draw_image(paste0(img_sur, 17, '.png'),   y = sdg_id.y[17], scale = scale_here) #+
  # draw_image(paste0(img_sur, 18, '.png'), y = 18-.5, scale = scale_here)

### insert the image strip into the bar plot and draw  
p_icon <- ggdraw(insert_yaxis_grob(plot = p1b, grob = icons, grid::unit(.1, "null"), position = "left"))
# p_icon

fig1b <- ggarrange(p_icon, labels = "b", font.label = list(size = 20), vjust = 1, hjust = 0)
fname <- paste0(dir.fig, '01main/fig1b_SDG_score_chg_Bar_Icon_', today, '.jpg'); fname
# ggsave(filename = fname, plot = fig1b, width = 6, height = 8, units = 'in', dpi = 600)

## ref: https://www.rdocumentation.org/packages/cowplot/versions/1.1.1/topics/axis_canvas
```

```{r ___ 17 SDG list}
```
SDG 2 (Zero hunger)
SDG 4 (Quality education)
SDG 6 (Clean water and sanitation)
SDG 11 (Sustainable cities and communities)
SDG 15 (Life on land) 




```{r - bar - change by Goal by 4 county group - fig 1c}
df_shp <- grp_update %>%
  merge(x = ., y = dfs_score_chg, by.x = 'iso_a3', by.y = 'iso3', all.x = T) %>%
  # st_drop_geometry() %>% 
  dplyr::mutate(
    group = tm::removeNumbers(group_income4),
    group = factor(x = group, levels = c("High income", "Upper middle income", "Lower middle income", "Low income")),
    group = fct_rev(group)) %>%
  as.data.frame()


chg_err_group  <- df_shp %>% 
  dplyr::filter(year != 'na') %>%
  summarySE(measurevar= "value_chg", groupvars = c("year", 'group', "goal"), na.rm = T)

# show_col(brewer_pal(palette = "Blues")(5))
color_grp <- brewer_pal(palette = "Blues")(5)[2:5]; color_grp
color_grp <- (color_grp) ## rev(color_grp)


ggplot(data = chg_err_group, 
       aes(x=factor(goal, levels = rev(vars)),
           y = value_chg,
           fill = group)) +
  geom_bar(position = position_dodge(0.8), stat="identity", width = 0.8, show.legend = T)+
  geom_errorbar(aes(ymin=value_chg-se, ymax=value_chg+se), 
                width=.3, 
                position=position_dodge(.8)) +
  ylab("Transnational impact on SDG score") + xlab('SDGs') + labs(fill = "Group")+
  # scale_fill_brewer(palette = "Blues", direction = -1) +
  scale_fill_manual(values = color_grp, guide = guide_legend(reverse = TRUE)) +
  coord_flip()+
  
  theme_bw() +
  theme(legend.position = c(0.15, 0.06), 
        # legend.key.size = unit(0.2,"cm"), 
        legend.key.width = unit(0.3,"cm"), 
        legend.key.height = unit(0.2,"cm"),
        panel.grid.minor = element_blank())

# fname <- paste0(dir.fig, '01main/SDG_score_chg_Bar_byGroup4_', today, '.jpg'); fname
# ggsave(filename = fname, plot = last_plot(), width = 6, height = 8, units = 'in', dpi = 300)

```




```{r - bar - change by Goal by 2 county group - fig 1c - !!}
## classify the country list into 2 groups
## -- option 1: high = high + upper middle; low = low + lower middle
mark <- 'option1'
shp_ctr_list <- grp_update %>% 
  dplyr::mutate(
    group = tm::removeNumbers(group_income4),
    group = ifelse(group %in% c("High income", "Upper middle income"), "High income", "Low income"),
    group = factor(x = group, levels = c("High income", "Low income")),
    group = fct_rev(group)
    ) %>%
  as.data.frame()

## -- option 2: high = high; low = low + lower middle + upper middle (Min Gon, Nature Food)
mark <- 'option2Chung'
shp_ctr_list <- grp_update %>% 
  dplyr::mutate(
    group = tm::removeNumbers(group_income2),
    group = factor(x = group, levels = c("High income", "Low income")),
    group = fct_rev(group)
    ) %>%
  as.data.frame()
```


```{r - change by group - MS ✏️📝}
## merge the two data for plot
df_shp <- shp_ctr_list %>%
  merge(x = ., 
        y = dfs_score_chg %>% dplyr::select(-group), 
        by.x = 'iso_a3', by.y = 'iso3', all.x = T) %>%
  as.data.frame()


chg_err_group  <- df_shp %>% 
  dplyr::filter(year != 'na') %>%
  summarySE(measurevar= "value_chg", groupvars = c("year", 'group', "goal"), na.rm = T)


chg_err_group_text <- chg_err_group %>%
  dplyr::select(year, group, goal, value_chg) %>%
  spread(key = group, value = value_chg) %>%
  mutate(good4who = ifelse(`High income` > `Low income`, 'better for High', 'better for Low')) 

chg_err_group_text %>%
  group_by(good4who) %>%
  tally()

cat('Low income - the number of SDGs that decreased:',  sum(chg_err_group_text$`Low income` < 0), '\n')
cat('High income - the number of SDGs that decreased:', sum(chg_err_group_text$`High income` < 0), '\n')

```


```{r ... fig 1c}
color_grp <- c('#2371A6', '#F26463') ## high, low
color_grp <- rev(color_grp)

p1c <-
  ggplot(data = chg_err_group, 
       aes(x=factor(goal, levels = rev(vars)),
           y = value_chg,
           fill = group)) +
  geom_bar(position = position_dodge(0.8), stat="identity", width = 0.8, show.legend = T)+
  geom_errorbar(aes(
    ymin=value_chg-se, ymax=value_chg+se), 
    width=.3, color = 'gray20',
    position=position_dodge(.8)) +
  ylab("Transnational impact on SDG score") + xlab('SDGs') + labs(fill = "Group")+
  # scale_fill_brewer(palette = "Blues", direction = -1) +
  scale_fill_manual(values = color_grp, guide = guide_legend(reverse = TRUE)) +
  scale_y_continuous(n.breaks = 6) +
  coord_flip()+
  
  theme_bar_my +
  theme(
    legend.position = c(0.877, 0.15),    ## 0.15, 0.06
    # legend.key.size = unit(0.2,"cm"), 
    legend.box.background = element_rect(colour = "black"),
    legend.key.width  = unit(0.4,"cm"), 
    legend.key.height = unit(0.4,"cm"), 
    legend.text=element_text(size=rel(.9)),
    panel.grid.minor = element_blank())

ggarrange(p1c, labels = "c", font.label = list(size = 20), vjust = 1)
fname <- paste0(dir.fig, '01main/fig1c_SDG_score_chg_Bar_byGroup2grp_', today, '_', mark, '.jpg'); fname
# ggsave(filename = fname, plot = last_plot(), width = 6, height = 8, units = 'in', dpi = 300)
```


```{r ... fig1bc combined-1}
p1bc <- ggarrange(fig1b, p1c+xlab(''), labels = "", font.label = list(size = 20), vjust = 1, hjust = 0, align = 'v')
fname <- paste0(dir.fig, '01main/fig1Combined1_bc_SDG_score_chg_Bar_Icon_', today, '.jpg'); fname
ggsave(filename = fname, plot = p1bc, width = 6*2, height = 8, units = 'in', dpi = 600)
```



```{r ... fig1bc combined-2}
chg_err_groups <- rbind(
  chg_err %>% dplyr::mutate(group = 'Global') %>% dplyr::select(names(chg_err_group)),
  chg_err_group
) %>%
  dplyr::mutate(group = factor(x = group, 
                               levels = c("Global", "High income", "Low income"), 
                               labels = c("Global", "High-income", "Low-income")),
                group = fct_rev(group))
  

# str(chg_err_groups)

color_grp <- c('gray80', '#2371A6', '#F26463') ## global, high, low
color_grp2 <- c('#2371A6', '#F26463') ## high, low
color_grp <- rev(color_grp)

p <- chg_err_groups %>%
  ggplot(data = ., 
         aes(x = factor(goal, levels = rev(vars)),
             y = value_chg,
             fill = group)) +
  geom_bar(position = position_dodge(0.8), stat="identity", width = 0.8, show.legend = T)+
  geom_errorbar(aes(ymin=value_chg-se, ymax=value_chg+se), 
                width=.3, 
                position=position_dodge(.8)) +
  ylab("Transnational impact on SDG score") + xlab('SDGs') + labs(fill = "Group")+
  # scale_fill_brewer(palette = "Blues", direction = -1) +
  scale_fill_manual(values = color_grp, guide = guide_legend(reverse = TRUE)) +
  scale_y_continuous(n.breaks = 6) +
  coord_flip()+
  
  theme_bar_my +
  theme(
    legend.position = c(0.877, 0.15),    ## 0.15, 0.06 -> 0.065
    # legend.key.size = unit(0.2,"cm"), 
    legend.box.background = element_rect(colour = "black"),
    legend.key.width  = unit(0.4,"cm"), 
    legend.key.height = unit(0.4,"cm"), 
    legend.text=element_text(size=rel(.9)),
    panel.grid.minor = element_blank())
p


### insert the image strip into the bar plot and draw  
p_icon <- ggdraw(insert_yaxis_grob(plot = p, grob = icons, grid::unit(.1, "null"), position = "left"))
# p_icon

fig <- ggarrange(p_icon, labels = "b", font.label = list(size = 20), vjust = 1, hjust = 0)
fname <- paste0(dir.fig, '01main/fig1Combined_bc_SDG_score_chg_Bar_Icon_', today, '.jpg'); fname
ggsave(filename = fname, plot = fig, width = 6, height = 8, units = 'in', dpi = 600)
```



```{r - map of 17 SDGs change - fig ex_3}

ind_sdg <- 'SDG Score change'

df_chg <- dfs_score_chg %>%
  dplyr::select(iso3, year, goal, value_chg) %>%
  spread(key = goal, value = value_chg)

###' need to define the break values and colors specifically -----------------------------
max(dfs_score_chg$value_chg, na.rm = T) ## 42.72
min(dfs_score_chg$value_chg, na.rm = T) ## -26.87468


### plot ---------------------------------------------------------------------------------
df_chg_sf  <- df_chg %>%
  merge(x = shp, y = ., by.x = "iso_a3", by.y = "iso3",  all.x = T)

col_tmap <- colors_changes
title_ls <- paste0('SDG ', vars); title_ls

fname <- paste0(dir.fig, '01main/ex_3_SDG_chg_Map_byGoal_', today, '.jpg'); fname
# jpeg(filename = fname, width=7, height=6, units="in", pointsize=18, res=500) ## width=7, height=4 --> ncol=4

col_tmap_score <- c(
  '#a63603','#e6550d','#fd8d3c',"#feb24c","#ffeda0", ## 6-class YlOrRd + 6-class Oranges
  '#eff3ff','#bdd7e7','#6baed6','#3182bd','#2171b5'  ## 6-class Blues
)

tm_shape(df_chg_sf) +
  tm_fill(vars, 
    palette = col_tmap_score,
    style = "fixed",
    breaks=c(-Inf, -20, -15, -10, -5, 0, 5, 10, 15, 20, Inf), 
    # midpoint = 0,
    labels = c("< -20", "[-20, -15)", "[-15, -10)", "[-10, -5]", "(-5, 0)", 
               "[0, 5)", "[5, 10]",   "(10, 15]",   "(15, 20]",  "> 20"), 
    title = title_ls, textNA = 'NA', colorNA = 'gray95',
    legend.reverse = T) +
  tm_borders(col = "grey", lwd = 0.1, lty = "solid", alpha = 0.99) +
  tm_layout(frame = T, frame.lwd = 0.1, 
    legend.position = c(0,0),
    legend.title.size = 0.9,
    legend.text.size  = 0.7,
    # legend.format = list(text.less.than = '<', text.or.more = "<", text.separator = "~"),
    legend.width = -0.7,
    legend.height = -0.7, 
    
    panel.show = F) +
  tm_facets(ncol=3)

# dev.off()
tmap_save(tm = tmap_last(), filename = fname, width=7, height=6, units="in", dpi = 300, outer.margins = 0)
```




```{r - map of the # of increased SDGs - fig 1d}

### for each country, how many goals improved from metacouplings? ------------------------
dfs_score_chg_count <- dfs_score_chg %>%
  # dplyr::filter() %>%
  dplyr::group_by(iso3) %>%
  dplyr::summarise(
    n_incr = sum(value_chg > 0, na.rm = TRUE),
    n_decr = sum(value_chg < 0, na.rm = TRUE),
    n = n(),
    n_na = sum(is.na(value_chg))) %>%
  ungroup() %>%
  dplyr::filter(n_na < 17/2) %>%  ## remove countries with too many NA
  dplyr::mutate(percent_increase = n_incr/17*100, 
                percent_decrease = 100 - percent_increase) %>%
  as.data.frame()

### plot ---------------------------------------------------------------------------------
df_chg_sf  <- dfs_score_chg_count %>%
  merge(x = shp, y = ., by.x = "iso_a3", by.y = "iso3",  all.x = T)

# col_tmap <- colors_changes
n_class = 10
n_class = 4
display.brewer.all()
col_tmap <- brewer.pal(n = n_class, name = "PiYG")
col_tmap <- brewer.pal(n = n_class, name = "Greens")

fname <- paste0(dir.fig, '01main/fig1d_SDG_score_count_Improved_', today, '_n', n_class, '.jpg'); fname
# jpeg(filename = fname, width=7, height=2.8, units="in", pointsize=18, res=300)

tm_shape(df_chg_sf) +
  tm_fill(col = 'percent_increase', 
          palette = col_tmap, #col1,
          style = "fixed",
          breaks= seq(0, 100, 100/n_class),
          title = '% of 17 SDGs\nimproved',
          textNA = 'NA', colorNA = 'gray90',
          legend.reverse = T) +
  tm_borders(col = "grey", lwd = 0.1, lty = "solid", alpha = 0.99) +
  tm_credits(text = "c", fontface = "bold", size = 1.2, position = c(0, 0.9)) +
  tm_layout(
    frame = F, frame.lwd = 0.1, 
    legend.position = c(0,0),
    legend.title.size = 0.8,
    legend.text.size  = 0.6,
    # legend.format = list(text.less.than = '<', text.or.more = "<", text.separator = "~"),
    legend.width = -0.7,
    legend.height = -0.7, 
    outer.margins=0, #c(0,0.001, 0, 0.001), #  bottom, left, top, and right
    inner.margins=0,
    panel.show = F) 
# dev.off()

tmap_save(tm = tmap_last(), filename = fname, width=7, height=2.8, units="in", dpi = 300)
```


```{r - map of the # of increased SDGs - fig 1d - % decreased}

### plot -------------------------------------------------------------------------------------------

col_tmap <- brewer.pal(n = n_class, name = "OrRd")

fname <- paste0(dir.fig, '01main/fig1dex_SDG_score_count_Decreased_', today, '_n', n_class, '.jpg'); fname
# jpeg(filename = fname, width=7, height=2.8, units="in", pointsize=18, res=300)

tm_shape(df_chg_sf) +
  tm_fill(col = 'percent_decrease', 
          palette = col_tmap, #col1,
          style = "fixed",
          breaks= seq(0, 100, 100/n_class),
          title = '% of 17 SDGs\ndecreased',
          textNA = 'NA', colorNA = 'gray90',
          legend.reverse = T) +
  tm_borders(col = "grey", lwd = 0.1, lty = "solid", alpha = 0.99) +
  tm_credits(text = "", fontface = "bold", size = 1.2, position = c(0, 0.9)) +
  tm_layout(
    frame = F, frame.lwd = 0.1, 
    legend.position = c(0,0),
    legend.title.size = 0.7,
    legend.text.size  = 0.6,
    # legend.format = list(text.less.than = '<', text.or.more = "<", text.separator = "~"),
    legend.width = -0.7,
    legend.height = -0.7, 
    outer.margins=0, #c(0,0.001, 0, 0.001), #  bottom, left, top, and right
    inner.margins=0,
    panel.show = F) 

# dev.off()

tmap_save(tm = tmap_last(), filename = fname, width=7, height=2.8, units="in", dpi = 300)
```



####  - Score by goal by scenario
```{r ... fig 3b}

dfs_score_goal_snr <- dfs_score %>%
  summarySE(measurevar= "score", groupvars = c("year", 'scenario', "goal"), na.rm = T) %>%
  as.data.frame() %>%
  dplyr::mutate_at(vars(score, sd, se, ci), list(~ round(., 2))) %>%
  ## for viz purpose, change `NaN` to a very small value
  dplyr::mutate_at(vars(score), ~ifelse(is.nan(.), 10e-9, .)) %>%
  dplyr::arrange(goal, scenario, year)


p <- dfs_score_goal_snr %>%
  ggplot(data = ., 
         aes(
           # x = goal, 
           x = fct_rev(goal),
           y = score,
           ymin = score-se, ymax = score+se,
           fill = fct_rev(scenario)
           )) +
  geom_bar(position = position_dodge(0.8), stat="identity", width = 0.8, show.legend = T, alpha = .9)+
  geom_errorbar(position=position_dodge(.9), width=.4, color = 'gray50') +
  scale_fill_manual(
    name   = 'Scenarios',
    values = rev(snr_colors), 
    labels = rev(snr_labels), 
    guide = guide_legend(reverse = TRUE)) +
  # Condition for labeling
  geom_text(position = position_dodge(0.8), 
            aes(label = case_when(
              score == 10e-9 ~ as.character('NA'), 
              score == 0 ~ paste0(' ', as.character(score)), 
              TRUE ~ "")),
            hjust = -0.5, size = 2.5) +
  xlab('SDGs') + ylab("Scores") +
  coord_flip() +
  theme_bar_my +
  theme(panel.grid = element_blank(),
        legend.position = c(0.872, 0.057), ## c(0.902, 0.057)
        legend.box.background = element_rect(colour = "black"),
        legend.key.width  = unit(0.4,"cm"),
        legend.key.height = unit(0.3,"cm"),
        axis.ticks.y = element_blank())
ggarrange(p, labels = "b", font.label = list(size = 20))
fname <- paste0(dir.fig, '01main/fig3b_score_goal_snr4_', today, '.jpg'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = 6, height = 8, units = 'in', dpi = 300)
```




####  - Score by goal by scenario by group

```{r ... fig ex_7}
dfs_score_goal_snr_grp <- dfs_score %>%
  summarySE(measurevar= "score", groupvars = c("year", 'group', 'scenario', "goal"), na.rm = T) 

p <-
  dfs_score_goal_snr_grp %>%
  dplyr::mutate(goal = paste0('SDG ', goal),
                goal = factor(x = goal, levels = paste0('SDG ', seq(1,17)) )) %>%
  ggplot(data = ., 
         aes(
           # x = goal, 
           # x = fct_rev(goal),
           x = scenario,
           y = score,
           ymin = score-se, ymax = score+se,
           fill = fct_rev(scenario),
           pattern = group
           )) +
  # geom_bar(position = position_dodge(0.8), stat="identity", width = 0.8, show.legend = T, alpha = .9)+
  geom_col_pattern(
    position = position_dodge(preserve = "single"),
    color = "gray30",
    # size  = 0.1,
    # pattern_fill = "gray50",
    # pattern_angle = 45,
    pattern_density = 0.1, ## 0.1
    pattern_spacing = 0.04,
    pattern_key_scale_factor = 0.6,
    alpha = .9) +
  scale_pattern_manual(
    name = 'Group', 
    values = c(`High income` = "circle", `Low income` = "none")) +
  geom_errorbar(position=position_dodge(.9), width=.4, color = 'gray30', linewidth = 0.3) +
  scale_fill_manual(
    name   = 'Scenarios',
    values = rev(snr_colors), 
    labels = rev(snr_labels), 
    guide = guide_legend(reverse = TRUE)) +
  # xlab('SDGs') + 
  xlab('Scenarios') +
  ylab("SDG scores") +
  # coord_flip() +
  scale_x_discrete(labels = snr_labels) +
  scale_y_continuous(limits = c(0, 100)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  facet_wrap(~goal, ncol = 4) + 
  theme_bar_my +
  theme(panel.grid = element_blank(),
        legend.position = c(1, -0.1), ## c(1, -0.05)
        legend.justification = c(1, 0), # right bottom justification
        legend.box.margin = margin(r = 0, b = 0, unit = "mm"), # small margin
        legend.box.background = element_rect(colour = "black"),
        legend.key.width  = unit(0.4,"cm"),
        legend.key.height = unit(0.3,"cm"),
        # legend.direction = "horizontal", 
        legend.box = "horizontal",
        axis.ticks.y = element_blank()) +
  guides(pattern = guide_legend(override.aes = list(fill = "white")), 
         fill    = guide_legend(override.aes = list(pattern = "none"), reverse = T) ) 
# p
fig <- ggarrange(p, labels = "", font.label = list(size = 20))
fname <- paste0(dir.fig, '01main/ex_7_score_goal_snr4_grp_', today, '.jpg'); fname
ggsave(filename = fname, plot = fig, width = 6, height = 8, units = 'in', dpi = 500)
```





##########################################################################################
### Index
##########################################################################################

```{r message=FALSE, warning=FALSE}
### SDG Index by country -----------------------------------------------------------------
dfs_index <- dfs_score %>%
  ungroup() %>%
  group_by(iso3, group, year, scenario) %>%
  dplyr::summarise(index = mean(score, na.rm = TRUE)) %>%
  ungroup() %>%
  dplyr::mutate(index = as.numeric(index)) %>%
  as.data.frame() 

### 
f <- paste0(dir.output, "SDG_index_byISO3_byYear.rds")
saveRDS(dfs_index, file = f)


### global SDG index
dfs_index_glo <- dfs_index %>%
  group_by(year, scenario) %>%
  dplyr::summarise(index = mean(index, na.rm = TRUE)) %>%
  ungroup() %>%
  column_to_rownames(var = 'scenario') %>%
  dplyr::mutate(scenario = row.names(.),
                scenario = factor(x = scenario, levels = snr_levels)) %>%
  as.data.frame()



### add global sdg index change to Fig 3a
snr_chg_levels <- c('value_chg_rel', 'value_chg_dst', 'value_chg_ner')

dfs_index_glo_chg <- dfs_index_glo %>%
  spread(key = scenario, value = index) %>%
  dplyr::mutate(
    value_chg_rel = value_rel - value_not, 
    value_chg_dst = value_dst - value_not, 
    value_chg_ner = value_ner - value_not, 
    ) %>%
  dplyr::filter(year == 2015) %>%
  dplyr::select(year, starts_with('value_chg_')) %>%
  gather(key = 'scenario', value = 'index', value_chg_rel:ncol(.)) %>%
  dplyr::mutate(group = 'Global') %>%
  summarySE(measurevar= "index", groupvars = c("year", 'scenario', 'group'), na.rm = T) %>%
  dplyr::mutate(scenario = factor(x = scenario, levels = snr_chg_levels)) %>%
  dplyr::arrange(year, scenario) %>%
  as.data.frame()



dfs_index_glo %>%
  ggplot(data = ., aes(x = scenario, y = index, fill = fct_rev(scenario))) +
  geom_col(width = .8, show.legend = F)+
  scale_fill_manual(
    name   = 'Scenarios',
    values = rev(snr_colors),
    labels = rev(snr_labels),
    guide  = guide_legend(reverse = T)) +
  scale_x_discrete(labels = snr_labels) +
  scale_y_continuous(limits = c(50, 60), breaks = seq(50, 60, 2), oob = rescale_none) +
  xlab('Scenarios') + ylab("Global SDG Index") +
  # theme_bar_my +
  theme(panel.grid = element_blank(),
        legend.position = c(0.873, 0.892), ## (0.894, 0.892)
        legend.box.background = element_rect(colour = "black"),
        legend.key.width = unit(0.4,"cm"),
        legend.key.height = unit(0.3,"cm"),
        ) 

```


#### - Index by scenario - bar
```{r ... fig3a}
dfs_index_snr <- dfs_index %>%
  summarySE(measurevar= "index", groupvars = c("year", 'scenario', 'group'), na.rm = T) %>%
  dplyr::arrange(year, scenario, group) %>%
  as.data.frame() 

y_upper <- ceiling(max(dfs_index_snr$index) + max(dfs_index_snr$se) + 1)
y_lower <- floor(min(dfs_index_snr$index) - max(dfs_index_snr$se) - 1)

### plot ------------------
p <- dfs_index_snr %>%
  ggplot(data = ., 
         aes(
           # x = group,
           x = scenario,
           y = index,
           ymin = index-se, ymax = index+se,
           fill = fct_rev(scenario),
           pattern = group
           )) +
  # geom_bar(position = position_dodge(0.8), stat="identity", width = 0.8, show.legend = T, alpha = .9)+
  
  geom_col_pattern(
    position = position_dodge(preserve = "single"),
    color = "gray30",
    # size  = 0.1,
    # pattern_fill = "gray50",
    # pattern_angle = 45,
    pattern_density = 0.1, ## 0.1
    pattern_spacing = 0.025,
    pattern_key_scale_factor = 0.6,
    alpha = .9) +
  
  scale_pattern_manual(
    name = 'Group', 
    values = c(`High income` = "circle", `Low income` = "none")) +
  
  scale_fill_manual(
    name   = 'Scenarios',
    values = rev(snr_colors), 
    labels = rev(snr_labels), 
    guide  = guide_legend(reverse = T)) +
  
  geom_errorbar(position=position_dodge(.9), width=.4, color = 'gray50') +
  scale_y_continuous(
    breaks = seq(y_lower, y_upper, 5),
    limits = c(y_lower, y_upper),
    oob = rescale_none
    ) +
  scale_x_discrete(labels = snr_labels) +
  xlab('Scenarios') + ylab("SDG Index") +
  # coord_flip() +
  theme_bar_my +
  theme(panel.grid = element_blank(),
        legend.position = c(0.873, 0.892), ## (0.894, 0.892)
        legend.box.background = element_rect(colour = "black"),
        legend.key.width = unit(0.4,"cm"),
        legend.key.height = unit(0.3,"cm"), 
        # legend.spacing.y = unit(0.0, 'cm'),
        # legend.margin = margin(t = 0),
        # axis.ticks.y = element_blank()
        ) +
  guides(pattern = guide_legend(override.aes = list(fill = "white")),
         fill    = guide_legend(override.aes = list(pattern = "none"), reverse = T) ) 
ggarrange(p, labels = "a", font.label = list(size = 20))
fname <- paste0(dir.fig, '01main/fig3a_index_byIncome_snr4_', today, '.jpg'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = 6, height = 8, units = 'in', dpi = 300)
```



#### - Index by scenario - map

```{r ... fig ex1}
dfs_index_sf <- dfs_index %>%
  spread(key = scenario, value = index) %>%
  merge(x = shp, y = ., by.x = "iso_a3", by.y = "iso3",  all.x = T)

hist(dfs_index_sf$value_rel, nclass = 50)
hist(dfs_index_sf$value_not, nclass = 50)
max_score <- (max(dfs_index_sf$value_rel, na.rm = T)/5) %>% ceiling() %>% "*"(5); max_score # 75
min_score <- (min(dfs_index_sf$value_rel, na.rm = T)/5) %>% floor()   %>% "*"(5); min_score # 35

### plot 
p <- tm_shape(dfs_index_sf) +
  tm_fill(c('value_rel', 'value_not'),
          palette = 'YlGn', 
          style = "cont",
          ## "cat", "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks"
          # stretch.palette = T,
          breaks = c(min_score, seq(50, max_score, by = 5), 100),
          # labels = seq(0, 100, by = 10),
          title  = c('SDG Index\n(Baseline)', 'SDG Index\n(Global autarky)'),
          textNA = 'NA', colorNA = 'gray90',
          legend.reverse = T) +
  tm_borders(col = "white", lwd = 0.1, lty = "solid", alpha = 0.99) +
  tm_layout(frame = F, frame.lwd = 0.1,
    legend.position = c(0,0),
    legend.title.size = 1.1, 
    # legend.title.fontface = "bold",
    legend.text.size  = 0.9,
    legend.width = -0.7,
    legend.height = -0.7,
    outer.margins=0, inner.margins=0,
    panel.show = F) +
  tm_facets(free.scales.fill = F, ncol = 1)
# p
fname <- paste0(dir.fig, '01main/', 'ex_1_SDG_Index_Map_Baseline_Lockdown_', today, '.jpg'); fname
tmap_save(tm = p, filename = fname, width=7, height=2.8*2, units="in", dpi = 300)
```




### Index change

```{r}

### SDG Index change ---------------------------------------------------------------------
dfs_index_chg <- dfs_index %>%
  # dplyr::select(-group) %>%
  spread(key = scenario, value = index) %>%
  dplyr::mutate(
    value_chg = value_rel - value_not, 
    # value_chg = ifelse(value_chg >= 20,  20, value_chg),
    # value_chg = ifelse(value_chg <=-20, -20, value_chg),
    value_chg = ifelse(iso3 =='GRL', NA, value_chg),
    chg_color = case_when(
      is.na(value_chg) ~ NA,
      value_chg < -5 ~ "#feb24c",
      value_chg <  0 ~ "#ffeda0",
      value_chg <  5 ~ '#eff3ff',
      value_chg < 10 ~ '#bdd7e7',
      value_chg < 15 ~ '#6baed6',
      value_chg < 20 ~ '#3182bd',
      T              ~ '#2171b5'
    )
    ) %>%
  dplyr::filter(year == 2015) %>%
  as.data.frame()


###' need to define the break values and colors specifically -----------------------------
break_seq_ini <- seq(-5, 20, 5)       

max(dfs_index_chg$value_chg, na.rm = T) ## 29.94641
min(dfs_index_chg$value_chg, na.rm = T) ## -5.548772
max_chg <- (max(dfs_index_chg$value_chg, na.rm = T)/5) %>% ceiling() %>% "*"(5); max_chg # 20   --> 30
min_chg <- (min(dfs_index_chg$value_chg, na.rm = T)/5) %>% floor()   %>% "*"(5); min_chg # -3.2 --> -10

c(
  "#feb24c","#ffeda0",
  '#eff3ff','#bdd7e7','#6baed6','#3182bd','#2171b5'
)


### % of change --------------------------------------------------------------------------
dfs_index_chg_pct <- dfs_index_chg %>%
  dplyr::filter(!is.na(value_not) & value_not != 0) %>%
  dplyr::mutate(value_chg_pct = value_chg/value_not*100)

mean_change <- mean(dfs_index_chg_pct$`value_chg_pct`, na.rm = T) %>% round(digits = 1)
sd_change <- sd(dfs_index_chg_pct$`value_chg_pct`, na.rm = T) %>% round(digits = 1)

hist(dfs_index_chg_pct$value_chg)
boxplot(dfs_index_chg_pct$value_chg)
summary(dfs_index_chg_pct$value_chg)
dfs_index_chg_pct %>%
  ggplot(aes(x=value_chg, fill = value_chg<0)) +
  geom_histogram(binwidth=1, na.rm = T, alpha = 0.7, show.legend = F) +
  scale_fill_manual(values = c('#3182bd', '#feb24c')) +
  geom_vline(aes(xintercept=0), color="red", linetype="dashed", linewidth=1) +
  # ggplot2::annotate("0", x = 0, y = 25, label = "") +
  geom_vline(aes(xintercept=mean_change), color="black", alpha=0.6, linewidth=1.5) +
  scale_x_continuous(breaks = seq(-15, 30, 5)) +
  ggplot2::annotate("text", x = mean_change+3, y = 20, label = paste0("mean = ", mean_change, '\nsd = ', sd_change)) +
  xlab('Difference of national SDG Index between the real-world scenario and an autarky scenario') + 
  ylab('Number of copuntries') +
  theme_bw()
```


```{r - MS ✏️📝}
### numbers for writing ------------------------------------------------------------------

##' ! Since global SDG Index is rarely used in literature, we won't include such a score. 
##'   G_Index = (Index_c1 + Index_c2 + … + Index_c189)/189
##'   G_Index* = (Index*_c1 + Index*_c2 + … + Index*_c189)/189
# index_glo_chg <- (dfs_index_glo['value_rel', 'index'] - dfs_index_glo['value_not', 'index']) %>% round(., digits = 1)
# index_glo_chg_pct <- (index_glo_chg/dfs_index_glo['value_not', 'index'] * 100) %>% round(., digits = 1)
# 
# cat('Global SDG Index changed', index_glo_chg, '(or', index_glo_chg_pct, '% ). \n\n')
# # Global SDG Index changed 4.2 (or 7.7 % ). 


cat('SDG Index of countries would improve', 
    mean(dfs_index_chg_pct$value_chg, na.rm = T) %>% round(., digits = 1),      ## 9.2 -> 4.1
    'on average.\n')
cat('SDG Index of countries would improve', 
    paste0(mean_change,  ## 20.2% --> 9.1%
    '% (SD = ', sd_change, '%) on average.\n'))

cat('SDG Index of countries in 2015:', 
    'mean =', 
    mean(dfs_index_chg_pct$value_rel, na.rm = T) %>% round(., digits = 1),  ## 20.2% --> 19.4%
    ', sd =', 
    sd(dfs_index_chg_pct$value_rel, na.rm = T) %>% round(., digits = 1),  ## 20.2% --> 19.4%
    '\n\n')


dfs_index_chg_count <- dfs_index_chg %>%
  merge(x=., 
        y=pop_l %>% dplyr::filter(year==2015) %>% dplyr::select(-year), 
        by.x='iso3', by.y='iso3', all.x=T) %>%
  dplyr::filter(!is.na(value_chg))  %>%
  dplyr::filter(!is.na(group)) %>%
  dplyr::mutate(increase     = ifelse(value_chg > 0, 1, 0),
                increase_pop = increase*pop)

## 1. number and % of country improved SDG index?
n_improved <- sum(dfs_index_chg_count$increase, na.rm = T) 
pct_improved <- (n_improved/nrow(dfs_index_chg_count)*100) %>%   ## 96.8%
   round(., digits = 0)
cat(pct_improved, '% (n=', n_improved, ') of the world countries improved their SDG Index.\n')

## 2. number and % of population improved SDG Index?
total_pop <- sum(dfs_index_chg_count$pop, na.rm = T); #total_pop
total_pop_increased <- sum(dfs_index_chg_count$increase_pop, na.rm = T); #total_pop_increased
(total_pop_increased/total_pop*100) %>%                             ## 98.5%
  round(., digits = 0) %>%
  cat('\t', ., '% of the world population improved their SDG Index.\n\n')



## 3. for countries with decreased SDG Index, how many are low-income and what it the % ?
cat('The remaining', 100-pct_improved, '% of the countries (n = ', 189-n_improved, 
    ') decreased their SDG Index.\n')
 
df_chg_sf_count_lowincome <- dfs_index_chg_count %>%
  dplyr::filter(increase < 1) %>%
  dplyr::filter(!is.na(group)) %>%
  dplyr::mutate(low_income = ifelse(
    stringr::str_detect(fixed(pattern = 'low', ignore_case = T), string = group), 1, 0)) %>%
  as.data.frame()
  
n_dec_low <- (sum(df_chg_sf_count_lowincome$low_income, na.rm = T)) 
(sum(df_chg_sf_count_lowincome$low_income, na.rm = T)/nrow(df_chg_sf_count_lowincome)*100) %>%
  round(., digits = 0) %>%
  cat('\t', ., '% (n=', n_dec_low, ') are low-income countries.')
## 76.5% are low-income
```




```{r - Index change - map}
dfs_index_chg_ <- dfs_index_chg %>%
  dplyr::select(-value_not, -value_rel)


### plot ---------------------------------------------------------------------------------
dfs_index_chg_sf  <- dfs_index_chg_ %>%
  merge(x = shp, y = ., by.x = "iso_a3", by.y = "iso3",  all.x = T) %>%
  dplyr::arrange(year, area, value_chg)


## - small regions that may be invisible on a map -----------------------
sf_small <- dfs_index_chg_sf %>%
  dplyr::mutate(area = as.numeric(area)/10^6) %>%
  dplyr::filter(area/10^4 < 3,
                !is.na(year))

sf_small_area <- sf_small %>%
  st_drop_geometry() %>%
  dplyr::select(iso_a3, area) %>%
  dplyr::distinct(iso_a3, .keep_all = T)

sf_small_decrease <- dfs_index_chg_sf %>%
  dplyr::mutate(area = as.numeric(area)/10^6) %>%
  dplyr::filter(value_chg < 0)
max(sf_small_decrease$area)

fname <- './Data/country classification/shp_prj_smaller3wkm2.shp'
# sf::st_write(sf_small, dsn = fname, delete_layer = T)

## in case when we want to make the small countries visible on the map, we can add a layer
## of point-line shape to highlight them 
### - smallest region to show
area_show_min <- 10^4
sf_small_p <- st_read('./Data/country classification/country_small_annotation_points.shp') %>%
  dplyr::select(-Country_Na) %>%
  merge(x = ., y = dfs_index_chg_, by.x = "ISO3", by.y = "iso3",  all.x = T) %>%
  merge(x = ., y = sf_small_area, by.x = "ISO3", by.y = "iso_a3",  all.x = T) %>%
  dplyr::filter(!is.na(year)) %>%
  dplyr::filter(area < area_show_min)


sf_small_l <- st_read('./Data/country classification/country_samll_annotation_lines.shp') %>%
  merge(x = ., y = dfs_index_chg_, by.x = "ISO3", by.y = "iso3",  all.x = T) %>%
  merge(x = ., y = sf_small_area, by.x = "ISO3", by.y = "iso_a3",  all.x = T) %>%
  dplyr::filter(!is.na(year)) %>%
  dplyr::filter(area < area_show_min)

## test plot
# tm_shape(dfs_index_chg_sf) +
#   tm_fill('value_chg') +
#   tm_shape(sf_small_l) +
#   tm_lines(lwd = .1, alpha = 0.7) +
#   tm_shape(sf_small_p) +
#   tm_dots(col = 'value_chg', size = 0.5)
## ----------------------------------------------------------------------

hist(dfs_index_chg_sf$value_chg)

# seq <- c(-20, -10, -5, 0, 5, 10, 20); seq; length(seq) ## should base on `min`
# col_tmap <- colors_changes ## Colour Palette; one value to one color ## ('red' --> 'blue')

break_seq <- seq(min_chg, max_chg, 5); break_seq; length(break_seq)
# col_tmap <- colors_changes[4:length(break_seq)]
col_tmap <- colors_changes[1:(length(break_seq))]
# col_tmap <- colors_changes
# col_tmap <- 'RdYlBu'

###' to make sure each categrory has the right color 
colors <- dfs_index_chg_sf %>% 
  dplyr::distinct(value_chg, chg_color) %>%
  arrange(value_chg) %>%
  dplyr::distinct(chg_color) %>%
  dplyr::filter(!is.na(chg_color))
col_tmap <- colors$chg_color
```


```{r ... fig 1a}
fname <- paste0(dir.fig, '01main/', 'fig1a_SDG_Index_chg_Map_', today, '.jpg'); fname
# jpeg(filename = fname, width=7, height=2.8, units="in", pointsize=18, res=300)

p <- tm_shape(dfs_index_chg_sf) +
  tm_fill(
    col = 'value_chg',
    stretch.palette = F, 
    # breaks = c(-Inf, min_chg, 0, 5, 10, 20, max_chg, Inf),
    
    ###' this option has to ensure the number of negative and positive breaks are the same
    ###'    so that the color can be symmetrical 
    # style  = "cont",
    # breaks = c(-30, -20, -15, -10, -5, 0, 5, 10, 15, 20, 30),
    # palette = 'RdYlBu',
    
    style  = "fixed",
    breaks = c(-Inf, break_seq_ini, Inf),
    palette = col_tmap, 
    legend.format = list(
      # scientific = TRUE, 
      # text.or.more = '<', text.less.than = '<',
      format="f", 
      text.separator = '–'),
    # labels = c("< -5", "[-5, 0)", "[0, 5)", "[5, 10)", "[10, 20]", paste0("> ", max_chg)),
    legend.reverse = T,
    textNA = 'NA', 
    colorNA = 'gray90',
    title  = 'Transnational\nimpact on\nSDG Index'
    ) +
  tm_borders(col = "grey", lwd = 0.1, lty = "solid", alpha = 0.99) +
  tm_credits(text = "a", fontface = "bold", size = 1.2, position = c(0, 0.9)) +
  
  ## to make small regions visible ------------- #
  tm_shape(sf_small_l) +
  tm_lines(lwd = .05, alpha = 0.5) +
  tm_shape(sf_small_p) +
  tm_dots(
    col = 'value_chg',
    palette = col_tmap, 
    stretch.palette = F,
    style  = "fixed",
    breaks = c(-Inf, break_seq_ini, Inf),
    # breaks = c(-5, 0, 5, 10, 20, Inf),
    # labels = c("(-5, 0)", "[0, 5)", "[5, 10]", "(10, 20]", "> 20"),
    legend.col.reverse = T, 
    legend.show = F,
    size = 0.02) +
  ## ------------------------------------------- #
  
  tm_layout(
    # main.title = 'a', main.title.position = c(0.1,1), 
    frame = F, frame.lwd = 0.1, 
    legend.position = c(0,0),
    legend.title.size = 0.9,
    legend.text.size  = 0.7,
    legend.width = -0.5,
    legend.height = -0.5,
    outer.margins=0, inner.margins=0, 
    panel.show = F)
p
# dev.off()
tmap_save(tm = p, filename = fname, width=7, height=2.8, units="in", dpi = 400)
```



```{r - change by 4 income group - boxplot}
```

```{r ... fig ex_2}
dfs_index_chg1 <- dfs_index_chg %>%
  merge(x = ., 
        y = grp_update, 
        by.x = 'iso3', by.y = 'iso_a3', all.x = T) %>%
  dplyr::mutate(
    group_income4 = factor(
      x = group_income4, 
      levels = c("High income", "Upper middle income", "Lower middle income", "Low income"))) %>%
  filter(!is.na(group)) 

dfs_index_chg1 %>%
  dplyr::filter(!is.na(value_chg)) %>%
  ggplot(aes(x= group_income4, y=value_chg, 
         # group = group_income4, 
         fill = group_income4, 
         color = group_income4)) + 
  geom_boxplot(outlier.shape = NA, notch = T, fill = NA, show.legend = F) + 
  geom_point(position=position_jitterdodge(jitter.width = 0.5),
             # aes(fill = group_income4), shape = 21, 
             # color = 'gray30',
             alpha = 0.6, show.legend = F) +
  # scale_color_brewer(palette = "Blues", direction = -1, na.value = 'black') +
  scale_colour_manual(values = brewer.pal(n = 6, "Blues")[6:3]) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  geom_hline(yintercept = 0, color="red", linetype="dashed") +
  xlab('') + ylab('Transnational impact on SDG Index')
  
fname <- paste0(dir.fig, '01main/ex_2_SDG_Index_chg_byGroup_4grp_', today, '.jpg'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = 6, height = 6, units = 'in', dpi = 300)
```


```{r - change by 2 income group - boxplot}
```


```{r ... fig ex_2}
dfs_index_chg %>%
  merge(x = ., 
        y = grp_update, 
        by.x = 'iso3', by.y = 'iso_a3', all.x = T) %>%
  dplyr::mutate(
    # group = ifelse(group %in% c("High income", "Upper middle income"), "High income", group),
    # group = ifelse(group %in% c("Lower middle income",  "Low income"), "Low income",  group),
    group = group_income2, 
    group = factor(x = group, levels = c("High income", "Low income"))
    ) %>%
  # dplyr::filter(!is.na(inc_group)) %>%
  as.data.frame() %>%
  
  ggplot(aes(x = group, y = value_chg, 
             group = group, fill = group, color = group)) + 
    geom_boxplot(outlier.shape = NA, notch = T, fill = NA, show.legend = F) + 
    geom_point(position=position_jitterdodge(jitter.width = 0.5),
               alpha = 0.6, show.legend = F) +
    scale_colour_manual(values = color_grp2) +
    # theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    geom_hline(yintercept = 0, color="red", linetype="dashed") +
    xlab('') + ylab('Transnational impact on SDG Index')
  
fname <- paste0(dir.fig, '01main/ex_2_SDG_Index_chg_byGroup_2grp_', today, '.jpg'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = 6, height = 6, units = 'in', dpi = 300)
```


```{r include=FALSE}
### free memory
gc()
```


#### - Change by scenario

```{r ... fig3a - chg}

dfs_index_chg_scr <- dfs_index %>%
  # dplyr::select(-group) %>%
  spread(key = scenario, value = index) %>%
  dplyr::mutate(
    value_chg_rel = value_rel - value_not, 
    value_chg_dst = value_dst - value_not, 
    value_chg_ner = value_ner - value_not, 
    # value_chg = ifelse(iso3 =='GRL', NA, value_chg),
    ) %>%
  dplyr::filter(year == 2015) %>%
  as.data.frame()

dfs_index_chg_scr_summ <- dfs_index_chg_scr %>%
  dplyr::select(iso3, group, year, starts_with('value_chg_')) %>%
  gather(key = 'scenario', value = 'index', 4:ncol(.)) %>%
  summarySE(measurevar= "index", groupvars = c("year", 'scenario', 'group'), na.rm = T) %>%
  dplyr::mutate(scenario = factor(x = scenario, levels = snr_chg_levels)) %>%
  dplyr::arrange(year, scenario, group) %>%
  ### add global index data
  rbind(dfs_index_glo_chg) %>%
  as.data.frame() 

y_upper <- ceiling(max(dfs_index_chg_scr_summ$index, na.rm = T) + max(dfs_index_chg_scr_summ$se, na.rm = T) + 1)
y_lower <- floor(min(dfs_index_chg_scr_summ$index, na.rm = T) - max(dfs_index_chg_scr_summ$se, na.rm = T))


str(dfs_index_chg_scr_summ)

n_grp <- unique(dfs_index_chg_scr_summ$group) %>% length()
if (n_grp == 3) {
  pattern_fill <- c('High income' = "circle", 'Low income' = "none", 'Global' = 'crosshatch')
  legend.y <- 0.891
} else {
  pattern_fill <- c('High income' = "circle", 'Low income' = "none")
  legend.y <- 0.899
}

### plot ------------------
p <- dfs_index_chg_scr_summ %>%
  ggplot(data = ., 
         aes(x = scenario, y = index,
             ymin = index-se, ymax = index+se,
             fill = fct_rev(scenario), pattern = group
           )) +
  # geom_bar(position = position_dodge(0.8), stat="identity", width = 0.8, show.legend = T, alpha = .9)+
  
  geom_col_pattern(
    position = position_dodge(preserve = "single"),
    color = "gray30",
    pattern_density = 0.1, ## 0.1
    pattern_spacing = 0.025,
    pattern_key_scale_factor = 0.6,
    alpha = .9) +
  
  scale_pattern_manual(
    name = 'Group', 
    values = pattern_fill) +
  
  scale_fill_manual(
    name   = 'Scenarios',
    values = rev(snr_colors[1:3]),
    labels = rev(snr_labels[1:3]),
    guide  = guide_legend(reverse = T)) +
  scale_x_discrete(labels = snr_labels[1:3]) +
  
  geom_errorbar(position=position_dodge(.9), width=.4, color = 'gray30') +
  scale_y_continuous(
    breaks = seq(y_lower, y_upper, 2),
    limits = c(y_lower, y_upper),
    oob = rescale_none
    ) +
  
  xlab('Scenarios') + ylab("SDG Index change") +
  # coord_flip() +
  theme_bar_my +
  theme(panel.grid = element_blank(),
        legend.position = c(0.875, legend.y), ## (0.894, 0.892)
        legend.box.background = element_rect(colour = "black"),
        legend.key.width = unit(0.4,"cm"),
        legend.key.height = unit(0.3,"cm"), 
        # legend.spacing.y = unit(0.0, 'cm'),
        # legend.margin = margin(t = 0),
        # axis.ticks.y = element_blank()
        ) +
  guides(pattern = guide_legend(override.aes = list(fill = "white")),
         fill    = guide_legend(override.aes = list(pattern = "none"), reverse = T) ) 
p
p3a <- ggarrange(p, labels = "a", font.label = list(size = 20))
fname <- paste0(dir.fig, '01main/fig3a_index_chg_snr_', today, '_addGlobal.jpg'); fname
ggsave(filename = fname, plot = p3a, width = 6, height = 8, units = 'in', dpi = 300)
```


### Indicator

  Map each SDG Indicator score under the two scenarios, i.e., baseline and lock down, for comparison. 
  
  *Note*: this won't be included in the manuscript, but is useful for data inspection or being used in SI. 
  
```{r - map scores, eval=FALSE, include=FALSE}

### 1. choose one indicator

str(dfs)

unique(dfs$sdg_ind)
unique(dfs$ind_sdg)

ind_sdg_i <- 'gdp_per_water'


df_score <- dfs %>%
  dplyr::mutate(
    goal     = factor(x = goal, levels = seq(1, 17, 1)),
    scenario = factor(x = scenario, levels = snr_levels)) %>%
  ungroup() %>%
  as.data.frame() %>%
  dplyr::filter(ind_sdg == ind_sdg_i)

ind_sdg_id <- df_score %>%
  distinct(sdg_ind) %>% unlist()
ind_sdg_id


### look at the data distribution ----------------------------------------
ggplot(df_score, aes(x = val)) +
  geom_histogram(bins = 30) +
  facet_wrap(~year) +
  theme_bw() +
  ggtitle(ind_sdg_i)

ggplot(df_score, aes(x = value_norm)) +
  geom_histogram(bins = 30) +
  facet_wrap(~year) +
  theme_bw() +
  ggtitle(ind_sdg_i)



### map plot eora ------------------------------------------------------------------------
df_sf  <- df_score %>%
  dplyr::select(-val) %>%
  # dplyr::filter(year == 2015) %>%
  merge(x = shp, y = ., by.x = "iso_a3", by.y = "iso3",  all.y = T) 

n_snr <- unique(df_sf$scenario) %>% length()

### any countries failed to be joined? --> ANT, SUN, SUN --> which are expected. 
df_sf_na <- df_sf %>%
  dplyr::filter(is.na(name))

map <- ggplot(data = df_sf) +
  geom_sf(aes(fill = value_norm), size = 0.1, color = "grey90") +
  scale_fill_viridis(direction = -1, na.value = "grey95") +
  facet_grid(year~scenario, switch = 'y')+ ## switch = 'y' ## --> label to be displaced on the left
  theme_bw() +
  theme_map
fname <- paste0(dir.fig, 'map_each_SDG_indicator_score/', 
                'map_SDG_2scenarios_', ind_sdg_id, '_', ind_sdg_i, '_', today, '.jpg'); fname
ggsave(filename = fname, plot = map, width = 7*n_snr/2, height = 5, units = 'in', dpi = 300)
```




```{r - map score change, eval=FALSE, include=FALSE}
df_score_chg_ <- df_score %>%
  dplyr::select(-val) %>%
  spread(key = scenario, value = value_norm) %>%
  dplyr::mutate(value_chg = value_rel - value_not, 
         # value_chg = ifelse(value_chg >= 20,  20, value_chg),
         # value_chg = ifelse(value_chg <=-20, -20, value_chg),
         value_chg = ifelse(iso3 =='GRL', NA, value_chg)) %>%
  dplyr::select(-value_not, -value_rel) %>%
  dplyr::filter(year == 2015) %>%
  as.data.frame()

hist(df_score_chg_$value_chg)



df_score_chg_sf  <- df_score_chg_ %>%
  merge(x = shp, y = ., by.x = "iso_a3", by.y = "iso3",  all.x = T)


seq <- c(-20, -10, -5, 0, 5, 10, 20); seq; length(seq)
labels.legend <- seq
# Colour Palette; one value to one color


### 1. using ggplot --> not that good
# ggplot(data = df_score_chg_sf) +
#   geom_sf(aes(fill = value_chg), size = 0.1, color = "grey90") +
#   # scale_fill_viridis(direction = -1, na.value = "grey20") +
#   scale_fill_gradient2(
#     low = "red", mid = "white", high = "blue",
#     midpoint = 0, space = "Lab", na.value = "grey95", guide = "colourbar",
#     aesthetics = "fill")+
#   # scale_fill_gradientn(colors=colors_changes, breaks=seq, na.value ='gray70', limits=c(-20, 20)) +
#   # guides(fill = guide_legend(label.hjust = 1, label = T))+
#   # facet_wrap(~goal)+
#   theme_bw() +
#   theme_map


### 2. using tmap --> good!
col_tmap <- colors_changes

# jpeg(filename = fname, width=7, height=4, units="in", pointsize=18, res=300)

tm_shape(df_score_chg_sf) +
  tm_fill(#sdgs_list,          # tm_polygons; tm_fill
    'value_chg',
          palette = col_tmap, #col1,
          style = "fixed",
          breaks=c(-Inf, -20, -10, -5, 0, 5, 10, 20, Inf),
          labels = c("< -20", "[-20, -10)", "[-10, -5]", "(-5, 0)", "[0, 5)", "[5, 10]", "(10, 20]", "> 20"), 
          title = ind_sdg , textNA = 'NA', colorNA = 'gray95',
          legend.reverse = T) +
  tm_borders(col = "grey90", lwd = 0.1, lty = "solid", alpha = 0.99) +
  tm_layout(frame = F, frame.lwd = 0.1,
    legend.position = c(0,0),
    legend.title.size = 0.7,
    legend.text.size  = 0.5,
    
    # legend.format = list(text.less.than = '<', text.or.more = "<",
    #                      text.separator = "~"),
    legend.width = -0.5,
    legend.height = -0.5,
    
    panel.show = F)

fname <- paste0(dir.fig, 'map_each_SDG_indicator_score/', 'map_SDG_score_chg_', ind_sdg_id, '_', ind_sdg_i, '_', today, '.jpg'); fname
tmap_save(tm = tmap_last(), filename = fname, width=7, height=4, units="in", dpi = 300)
```





```{r - map compare with WIOD, eval=FALSE, include=FALSE}
### map plot wiod ----------------------------------------------------------------------------------
### for comparison  
pattern <- paste0('normalized_score__', ind_sdg , '.RData'); pattern
fname <- list.files(path = 'G:/My Drive/_paper/SDGs/SDGs_Global_Trade', 
                    recursive = T, full.names = T,
                    pattern = pattern); fname
# length(fname) 
# nchar(fname)

if (length(fname) > 0) {
  load(fname)
  df_wiod <- normalized_score %>%
    ### format data
    dplyr::filter(scenario %in% c('rel_c', 'not_c')) %>%
    dplyr::select(nation, scenario, '2000', '2005', '2009') %>% 
    gather(key = 'year', value = 'value_norm',   `2000`:ncol(.)) %>%
    as.data.frame() %>%
    ### join with shp
    merge(x = shp_wiod, y = ., by.x = "iso_a3", by.y = "nation",  all.y = T) %>%
    ### We remove the ROW data here for better comparison with the Eora data. 
    dplyr::mutate(value_norm = ifelse(iso_a3 == 'ROW', NA, value_norm))
  
  # unique(df_wiod$iso_a3)
  # names(df_wiod)
  
  ggplot(data = df_wiod) +
    geom_sf(aes(fill = value_norm), size = 0.1, color = "grey90") +
    scale_fill_viridis(direction = -1, na.value = "grey95") +
    facet_grid(year~scenario, switch = 'y')+ ## switch = 'y' ## --> label to be displaced on the left
    theme_bw() +
    theme_map  
  fname <- paste0(dir.fig, 'map_each_SDG_indicator_score/', 'map_SDG_2scenarios_', 
                  ind_sdg_id, '_', ind_sdg_i, '_wiod_', today, '.jpg'); fname
  ggsave(filename = fname, plot = last_plot(), width = 7, height = 4, units = 'in', dpi = 300)
} else {
  print('No such data')
}

```

















# ========================================================================================


# MC: distant vs. nearby

  *How to determine?*
  
  1. Boolean (0 or 1) to determine Distant vs. nearby matrix
  
  
    |      | ctr1 | ctr2 | ctr3  | ... |
    --------------------------------------
    | ctr1 |  1   |   1  |    0  |     |
    | ctr2 |  1   |   1  |    0  |     |
    | ctr3 |  0   |   0  |    0  |     |
    | ...  |      |      |       |     |
    
    
  2. Quantitative euclidean distance to determine Distant vs. nearby matrix 
  
  
    |      | ctr1 | ctr2 | ctr3  | ... |
    --------------------------------------
    | ctr1 |    1 | 100  |  900  |     |
    | ctr2 |  100 |   1  |  300  |     |
    | ctr3 |  900 | 300  |    1  |     |
    | ...  |      |      |       |     |
    
    Then, if
        d < 100,    nearby,  truck
        d < 1000,   distant, train/ship
        else,       distant, flight?
    
```{r eval=FALSE, include=FALSE}
Net_Import_Cal <- function(df.xlsx, matrix) {
  # trade matrix
  s1 <- read_excel(df.xlsx, sheet = 1, col_names = T, range = "B1:AP616")
  s2 <- matrix
  
  # nearby trade matrix
  s3 <- s1 * s2
  
  # add country name as a new col
  countries <- colnames(s2[1:41])
  colnames(s3) <- countries; length(countries)
  
  # add year seq to data frame 
  year_seq = rep(seq(1995, 2009, by=1), each = 41)
  s3$year <- year_seq
  
  # total nearby imports
  s4 <- s3 %>% group_by(year) %>% summarise_all(funs(sum))
  
  # tanspose data frame and keep the first col as new header
  s5 = as.data.frame(t(s4[,-1]))
  colnames(s5) <- seq(1995, 2009, by=1)
  s5 <- s5[ order(row.names(s5)), ]
  
  # total nearby exports
  library("reshape2") # expand long table to wide one
  s6 <- s3 %>%
    mutate(row.sum   = rowSums(s3[,1:41]),
           year      = year_seq,
           countries = rep(countries, times = 15)) %>% # times=15, each = 15
    dcast(countries ~ year, value.var = 'row.sum') # year as new col names
  s6 <- data.frame(s6[,-1], row.names=s6[,1])
  s6 <- s6[ order(row.names(s6)), ]
  
  # net nearby imports
  s7 <- s5 - s6
}


# load spatial relation: nearby or distant
ner_dst_matrix <- "./_input_data/_adjancet_distant_matrix.xlsx"
ner.matrix <- read_excel(path = ner_dst_matrix, sheet = 'S2.country_nearby',  col_names = T, range = "B1:AP616")
dst.matrix <- read_excel(path = ner_dst_matrix, sheet = 'S2.country_distant', col_names = T, range = "B1:AP616")
not.matrix <- read_excel(path = ner_dst_matrix, sheet = 'S2.country_not',     col_names = T, range = "B1:AP616")

# cal using function
s7.ner <- Net_Import_Cal(df.xlsx, matrix = ner.matrix)
s7.dst <- Net_Import_Cal(df.xlsx, matrix = dst.matrix)
s7.not <- Net_Import_Cal(df.xlsx, matrix = not.matrix)
```
