==> Code/27b_TNI_Networks_map.Rmd <==
---
title: "Untitled"
author: "Yingjie"
date: "6/30/2021"
output: html_document
editor_options: 
  chunk_output_type: inline
---

# Overview

  To get the output for the analysis, we only need to run these two chunks
  - `#### --> All indicators`
  - `# 4. Plot network for each TNI`
  
  Other sections/chunks are for testing and selecting the best viz option. 
  


## 11. Map network

Plot a network graph of nodes with geographic coordinates on a map.
Author: Markus Konrad <markus.konrad@wzb.eu>
May 2018
https://www.r-bloggers.com/2018/05/three-ways-of-visualizing-a-graph-on-a-map/
https://gist.github.com/internaut/a9a274c72181eaa7f5c3ab3a5f54b996#file-networkmap-r

This script shows three ways of plotting a network graph on a map.
The following information should be visualized (with the respective aestethics added):

  * graph nodes with:
    * position on map -> x,y position of the node
    * node weight (degree of the node) -> node size
    * node label -> also x,y position of the node
  * edges between nodes with:
    * edge weight -> edge width
    * edge category -> edge color
  
```{r}

### To clear your environment 
# remove(list = ls())

### set work dir
# path <- rstudioapi::getSourceEditorContext()$path
# dir  <- dirname(rstudioapi::getSourceEditorContext()$path); dir
# setwd(dir)
# setwd('..') # set directory by one folder up
getwd()

### data path
source('./Code/_path of data.R')

library(assertthat)
library(dplyr)
library(purrr)
library(igraph)
library(ggplot2)
library(ggraph)
library(ggmap)
library(ggpubr)

today <- format(Sys.time(), "%Y%m%d"); today
```


```{r - data example}
# -------------------------------------- #
# Preparation: generate some random data #
# -------------------------------------- #

country_coords_txt <- "
   1     3.00000  28.00000       Algeria
   2    54.00000  24.00000           UAE
   3   139.75309  35.68536         Japan
   4    45.00000  25.00000 'Saudi Arabia'
   5    9.00000   34.00000       Tunisia
   6     5.75000  52.50000   Netherlands
   7   103.80000   1.36667     Singapore
   8   124.10000  -8.36667         Korea
   9    -2.69531  54.75844            UK
  10    34.91155  39.05901        Turkey
  11  -113.64258  60.10867        Canada
  12    77.00000  20.00000         India
  13    25.00000  46.00000       Romania
  14   135.00000 -25.00000     Australia
  15    10.00000  62.00000        Norway"

# nodes come from the above table and contain geo-coordinates for some
# randomly picked countries
nodes <- read.delim(text = country_coords_txt, header = FALSE,
                    quote = "'", sep = "",
                    col.names = c('id', 'lon', 'lat', 'name')) %>%
  dplyr::rename('label' = 'name')

# edges: create random connections between countries (nodes)
set.seed(123)

N_EDGES_PER_NODE_MIN <- 1
N_EDGES_PER_NODE_MAX <- 4
N_CATEGORIES <- 4

edges <- map_dfr(nodes$id, function(id) {
  n <- floor(runif(1, N_EDGES_PER_NODE_MIN, N_EDGES_PER_NODE_MAX+1))
  to <- sample(1:max(nodes$id), n, replace = FALSE)
  to <- to[to != id]
  categories <- sample(1:N_CATEGORIES, length(to), replace = TRUE)
  weights <- runif(length(to))
  data_frame(from = id, to = to, weight = weights, category = categories)
})

edges <- edges %>% mutate(category = as.factor(category))


nodes.example <- nodes
edges.example <- edges
```


**...Tips...** 
edges
  - weight to be normalized to 0-1
  

```{r - your data - simplify}

n_flows_to_show <- 200 ## 105

load('./Data/Ancillary_Data_ISO3code_shp.RData') ## `iso_eora`, `pop`, `gdp` (USD), `shp`, `grp`, `grp_update`
rm(pop, gdp, grp)


###' the data were derived from `27a_TNI_Networks.Rmd`
###'  - The countries in the `to` represents who impose negative impacts on others
f <- paste0('./Data/data_02_intermediate/dt02_flows/', 'network_data_for_viz.RData')
load(f) ## `links_scaled_all_Net`, `grp_update_net`, `pop`


### load capital city data
f <- "./Data/capital_city_data.rds"
capital_city_data <- readRDS(file = f) %>%
  dplyr::rename('lon' = 'long') %>%
  dplyr::mutate(
    lon = case_when(
      iso3=='AUS' ~ lon-15,
      iso3=='CAN' ~ lon-35,
      iso3=='CHN' ~ lon-5,
      iso3=='DEU' ~ lon-3,
      iso3=='GBR' ~ lon-1,
      iso3=='FIN' ~ lon+1,
      iso3=='MEX' ~ lon-5,
      iso3=='RUS' ~ lon+65,
      iso3=='SWE' ~ lon-1,
      iso3=='USA' ~ lon-20,
      T           ~ lon),
    lat = case_when(
      iso3=='AUS' ~ lat+10,
      iso3=='CHN' ~ lat-5,
      iso3=='CAN' ~ lat+15,
      iso3=='DEU' ~ lat-2,
      iso3=='GBR' ~ lat+1,
      iso3=='FIN' ~ lat+5,
      iso3=='IND' ~ lat-15,
      iso3=='RUS' ~ lat+5,
      iso3=='SWE' ~ lat+5,
      T           ~ lat)
  ) %>%
  dplyr::select(iso3, lat, lon)

### choose one of the indicators to compare 
ind <- 'Landuse_forest'
ind <- 'total'

  
source('https://raw.githubusercontent.com/Yingjie4Science/functionX/main/scaling_normalization_Pro.R')

edges <- links_scaled_all_Net %>%
  dplyr::select(all_of(c('from', 'to', ind))) %>%
  ##' need to scale the range of weight to 0-1
  dplyr::mutate(weight = func_norm_minmax(x=total, a=0, b=4), 
                weight = ifelse(weight<0, 0, weight)) %>%
  ##' look at the top 15 flows
  arrange(desc(weight)) %>%
  head(n_flows_to_show) %>%
  ## remove nodes with too few links (< 5 links)
  group_by(from) %>%
  dplyr::mutate(freq_from = n()) %>%
  dplyr::filter(freq_from >= 1) %>%  
  dplyr::select(-freq_from) %>%
  ungroup() %>%
  as.data.frame()


nodes <-
  data.frame(iso3 = c(edges$from, edges$to)) %>%
  dplyr::mutate(iso3 = as.character(iso3)) %>%
  dplyr::distinct(iso3, .keep_all = T) %>%
  merge(x = ., 
        y = grp_update_net, 
        by.x = 'iso3', by.y = 'iso_a3', all.x = T) %>%
  dplyr::distinct(iso3, .keep_all = T) %>%
  dplyr::mutate(group = as.factor(group_income2)) %>%
  # merge(x = ., 
  #       y = pop %>% dplyr::select(iso3, `2015`), 
  #       by.x = 'iso3', by.y = 'iso3', all.x = T) %>%
  # dplyr::rename('size_node' = '2015') %>%
  arrange(iso3) %>%
  # dplyr::mutate(node_type  = ifelse(group == 'High income', 1, 0), 
  #               type.label = group) %>%
  dplyr::mutate(id    = iso3, 
                label = iso3) %>%
  ### add capital city's lon and lat info
  left_join(., capital_city_data, by = c('iso3')) %>%
  # dplyr::distinct(id, .keep_all = T) %>%
  dplyr::select(id, iso3, everything()) %>%
  as.data.frame()

# Remove duplicates based on Sepal.Width columns
nodes[duplicated(nodes$id), ]
```



```{r - plot}
# create the igraph graph object

g <- graph_from_data_frame(edges, directed = T, vertices = nodes)

# --------------------------------------------------------------------- #
# Common data structures and ggplot objects for all the following plots #
# --------------------------------------------------------------------- #

# create a data frame for plotting the edges
# join with nodes to get start and end positions for each
# edge (x, y and xend, yend)

edges_for_plot <- edges %>%
  ### keep the top n edges
  arrange(desc(weight)) %>%
  head(n_flows_to_show) %>%
  inner_join(nodes %>% dplyr::select(id, lon, lat), by = c('from' = 'id')) %>%
  rename(x = lon, y = lat) %>%
  inner_join(nodes %>% dplyr::select(id, lon, lat, group), by = c('to' = 'id')) %>%
  rename(xend = lon, yend = lat) %>%
    dplyr::mutate(category = group)
# assert_that(nrow(edges_for_plot) == nrow(edges))

# use the node degree for scaling the node sizes
nodes_for_plot <- nodes %>%
  dplyr::mutate(weight = igraph::degree(g, mode="in")) %>%
  dplyr::filter(id %in% c(edges_for_plot$from, edges_for_plot$to))



# common polygon geom for plotting the country shapes
country_shapes <- geom_polygon(
  data = map_data('world') %>% dplyr::filter(region != 'Antarctic'), 
  aes(x = long, y = lat, group = group),
  # fill = "transparent", 
  fill = "gray90", 
  color = "gray70", size = 0.05, alpha = .5)

dd <- map_data('world')
# country_shapes <- geom_sf(
#   data = shp, 
#   fill = "transparent", color = "gray50", size = 0.05, alpha = .5)



# common coordinate system for all the following plots
mapcoords    <- coord_fixed(xlim = c(-150, 180), ylim = c(-55, 80))
mapcoords_eu <- coord_fixed(xlim = c(-10, 36),   ylim = c(34, 65), clip = 'on')


color_label <- "black"
color_label <- "transparent"

# common plot theme
maptheme <- theme(
  panel.border = element_blank(),
  panel.grid = element_blank()) + 
  theme(axis.text = element_blank()) +
  theme(axis.ticks = element_blank()) +
  theme(axis.title = element_blank()) +
  theme(#legend.position = "bottom", 
        legend.background = element_rect(fill = "transparent")) +
  theme(panel.grid = element_blank()) +
  theme(panel.background = element_rect(fill = "transparent")) +
  theme(plot.margin = unit(c(0, 0, 0.5, 0), 'cm')) ## t, r, b, l



unique(nodes_for_plot$group)
n_grp <- unique(nodes_for_plot$group) %>% length()
source('./Code/helper_color_scheme.R')
if (n_grp > 2) {
  fill_color <- c('High income'         = color_income_group_4[1],
                  'Upper middle income' = color_income_group_4[2],
                  'Lower middle income' = color_income_group_4[3],
                  'Low income'          = color_income_group_4[4])
  label_edge <- c('Impact by High income', 
                  'Impact by Upper middle income',
                  'Impact by Lower middle income',
                  'Impact by Low income' 
                  )
} else {
  fill_color <- c('High income' = color_income_group_2[1], 
                  'Low income'  = color_income_group_2[2])
  
  label_edge <- c('Impact by High income', 
                  'Impact by Low income' 
                  )
}
```



### Solution 1: ggplot + ggmap only
```{r}
# ------------------------------- #
# Solution 1: ggplot + ggmap only # -------------
# ------------------------------- #

# try to plot with scaled edge widths and node sizes
# this will fail because we can only use the "size" aesthetic twice
# ggplot(nodes_for_plot) + 
#   country_shapes +
#   geom_curve(aes(x = x, y = y, xend = xend, yend = yend,     # draw edges as arcs
#                  color = category, size = weight),
#              arrow = arrow(length = unit(0.2, "in"), ends = 'last', 
#                            angle = '20', # default = 30
#                            type = 'open'),
#              data = edges_for_plot, curvature = 0.33, alpha = 0.5) +
#   scale_size_continuous(guide = FALSE, range = c(0.25, 2)) + # scale for edge widths
#   geom_point(aes(x = lon, y = lat, size = weight),           # draw nodes
#              shape = 21,   
#              fill = 'white', color = 'black', stroke = 0.5) +
#   scale_size_continuous(guide = FALSE, range = c(1, 6)) +    # scale for node size
#   geom_text(aes(x = lon, y = lat, label = label),             # draw text labels
#             hjust = 0, nudge_x = 1, nudge_y = 4,
#             size = 3, color = color_label) +
#   mapcoords + maptheme



##' Results in warning: "Scale for 'size' is already present. 
##'   Adding another scale for 'size', which will replace the existing scale."
##' now a plot with static node size:

ggplot(nodes_for_plot) + 
  country_shapes +
  geom_curve(
    data = edges_for_plot %>% head(100), 
    aes(x = x, y = y, xend = xend, yend = yend,     # draw edges as arcs
                 color = category, 
                 alpha = weight,
                 size = weight),
             # arrow = arrow(length = unit(0.1, "in"), ends = 'last', 
             #               angle = '20', # default = 30
             #               type = 'open'),
             arrow = arrow(length = unit(0.04, "inches"), ends = 'first'), ## One of "last" , "first" , or "both"
             lineend = 'butt', show.legend = F,
             curvature = 0.33) +
  scale_size_continuous(guide = FALSE, range = c(0.1, .8)) + # scale for edge widths
  scale_alpha_continuous(range = c(0.2, 1)) +
  scale_color_manual(values = fill_color) +
  geom_point(aes(x = lon, y = lat,
                 fill = group),                          # draw nodes
             shape = 21, size = 1.5, alpha = .7, show.legend = F,
             # fill = 'white', 
             color = 'black', stroke = 0.1) +
  # scale_size_continuous(guide = FALSE, range = c(.5, 1.5)) +    # scale for node size
  scale_fill_manual(values = fill_color) +
  geom_text(aes(x = lon, y = lat, label = label),             # draw text labels
            hjust = 0, nudge_x = .5, nudge_y = .5,
            size = 1, color = 'black') + # , fontface = "bold"
  # mapcoords + 
  mapcoords_eu +
  maptheme +
  theme(legend.position = c(.1,0))

fname <- paste0(dir.fig, '01main/fig2c_network_', n_flows_to_show, '_', today, 'eu2.jpg'); fname
ggsave(filename = fname, plot = ggplot2::last_plot(), width = 3, height = 2, units = 'in', dpi = 600)
```


### Solution 2: ggplot2 + ggmap + ggraph
```{r eval=FALSE, include=FALSE}
# ------------------------------------ #
# Solution 2: ggplot2 + ggmap + ggraph #
# ------------------------------------ #

# prepare layout: use "manual" layout with geo-coordinates
node_pos <- nodes_for_plot %>% select(lon, lat) %>% rename(x = lon, y = lat)
lay <- ggraph::create_layout(graph = g, layout = 'manual', node.positions = node_pos)
assert_that(nrow(lay) == nrow(nodes_for_plot))

# use the node degree for scaling the node sizes
lay$weight <- igraph::degree(g)

ggraph(lay) + country_shapes +
  geom_edge_arc(aes(color = category, edge_width = weight,   # draw edges as arcs
                    circular = FALSE),
                data = edges_for_plot, curvature = 0.33, alpha = 0.5) +
  scale_edge_width_continuous(range = c(0.5, 2),             # scale for edge widths
                              guide = FALSE) +
  geom_node_point(aes(size = weight), shape = 21,            # draw nodes
                  fill = "white", color = "black",
                  stroke = 0.5) +
  scale_size_continuous(range = c(1, 6), guide = FALSE) +    # scale for node widths
  geom_node_text(aes(label = name), repel = TRUE, size = 3,
                 color = color_label, fontface = "bold") +
  mapcoords + maptheme
```



### Solution 3: the hacky way (overlay several ggplot "plot grobs")
```{r}
# --------------------------------------------------------------- #
# Solution 3: the hacky way (overlay several ggplot "plot grobs") #
# --------------------------------------------------------------- #

test_postfix <- '7'

theme_transp_overlay <- theme(
  panel.background = element_rect(fill = "transparent", color = NA),
  plot.background = element_rect(fill = "transparent", color = NA), 
  panel.border = element_blank(),
) #+
  # scale_y_continuous(expand = c(0, 0)) + ## reduce the space between plot and border
  # scale_x_continuous(expand = c(0, 0)) ## reduce the space between plot and border



# the base plot showing only the world map
p_base <- ggplot() + country_shapes + mapcoords + maptheme

# first overlay: edges as arcs
p_edges <- ggplot(edges_for_plot) +
  geom_curve(aes(x = x, y = y, xend = xend, yend = yend,     # draw edges as arcs
                 color = category, 
                 alpha = weight,
                 size = weight),
             curvature = 0.45, ## 0.33
             # alpha = 0.5,
             arrow = arrow(length = unit(0.06, "inches"), ends = 'first'), ## One of "last" , "first" , or "both"
             lineend = 'butt', 
             show.legend = T,
             ) +
  scale_size_continuous(guide = FALSE, range = c(0.1, .8)) +  # scale for edge widths
  scale_alpha_continuous(range = c(0.2, .8)) +
  scale_color_manual(values = fill_color, labels = label_edge, name = 'Impact direction') +
  mapcoords + maptheme + theme_transp_overlay +
  theme(
    legend.position = c(0, 0), 
    legend.justification = c(0, 0), # left bottom justification
    legend.box.margin = margin(r = 0, b = 0, unit = "mm"), # small margin)
    legend.background = element_rect(fill = "transparent", color = NA),
  ) +
  guides(alpha = "none")  ## , size = "none"

# second overlay: nodes as points
p_nodes <- ggplot(nodes_for_plot) +
  geom_point(aes(x = lon, y = lat, 
                 fill = group,
                 size = weight
                 ),
             shape = 21, show.legend = T,
             # fill = "white",
             color = "black",    # draw nodes
             # size = 1.5, 
             alpha = .7,
             stroke = 0.1) + # stroke = 0.1
  scale_size_continuous(guide = F, ## to show legend or not
                        range = c(0.5, 1.6), ) +    # scale for node size
  scale_fill_manual(values = fill_color, name = 'Country group') +
  geom_text(aes(x = lon, y = lat, label = label),             # draw text labels
            hjust = 0, nudge_x = 1, nudge_y = 4,
            size = 3, color = color_label) +
  mapcoords + maptheme + theme_transp_overlay +
  theme(
    legend.position = c(0, 0.3), 
    legend.justification = c(0, 0), # left bottom justification
    legend.box.margin = margin(r = 0, b = 0, unit = "mm"), # small margin)
    legend.background = element_rect(fill = "transparent", color = NA),
  )

# combine the overlays to a full plot

# proper positioning of the grobs can be tedious... I found that
# using `ymin` works quite well but manual tweeking of the
# parameter seems necessary
ymin_my <- -74 ## -74
p <- p_base +
  annotation_custom(ggplotGrob(p_edges), ymin = ymin_my) +
  annotation_custom(ggplotGrob(p_nodes), ymin = ymin_my) 

# print(p)

(p2c <- ggarrange(p, labels = "c"))
fname <- paste0(dir.fig, '01main/fig2c_network_', n_flows_to_show, '_', today, test_postfix,  '.jpg'); fname
ggsave(filename = fname, plot = p2c, width = 7, height = 3.2, units = 'in', dpi = 600)
```




  Test code - not working
```{r eval=FALSE, include=FALSE}
library(ggforce)
#> Loading required package: ggplot2
#> 

ymin_my2 <- 35

p_base2 <- ggplot() + country_shapes + mapcoords_eu + maptheme


p_base2 + 
  annotation_custom(ggplotGrob(p_edges), xmin = -105, xmax = 100, ymin = ymin_my2) +
  annotation_custom(ggplotGrob(p_nodes), xmin = -105, xmax = 100, ymin = ymin_my2)
```
